% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_data_sim.R
\name{add_cat}
\alias{add_cat}
\title{Add organizational data to the simulated p2p data}
\usage{
add_cat(x, type)
}
\arguments{
\item{x}{Numeric vector to be assigned simulated organizational attributes}

\item{type}{String with three valid options:
\itemize{
\item \code{Organization}
\item \code{LevelDesignation}
\item \code{City}
}}
}
\description{
Add organizational data to the simulated p2p data
}
