% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_fizz.R
\name{afterhours_fizz}
\alias{afterhours_fizz}
\title{Distribution of After-hours Collaboration Hours (Fizzy Drink plot)}
\usage{
afterhours_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze weekly after-hours collaboration hours distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
}
\examples{
afterhours_fizz(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Collaboration: 
\code{\link{collaboration_area}()},
\code{\link{collaboration_dist}()},
\code{\link{collaboration_fizz}()},
\code{\link{collaboration_line}()},
\code{\link{collaboration_rank}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()},
\code{\link{create_dist}()},
\code{\link{meeting_trend}()}
}
\concept{Collaboration}
