% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_rank.R
\name{afterhours_rank}
\alias{afterhours_rank}
\title{Rank groups with high After-Hours Collaboration Hours}
\usage{
afterhours_rank(data, hrvar = extract_hr(data), mingroup = 5, return = "table")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{A character vector specifying what to return.
Valid values include "html" (default, returning an interactive DataTable)
and "df" (data frame)}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
This function scans a Standard Person Query for groups with high levels of After-Hours Collaboration.
Returns a table with a all of groups (across multiple HR attributes) ranked by hours of After-Hours Collaboration Hours.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
}
\seealso{
Other After-Hours: 
\code{\link{afterhours_dist}()},
\code{\link{afterhours_summary}()},
\code{\link{afterhours_trend}()}
}
\concept{After-Hours}
