% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_csv.R
\name{bind_csv}
\alias{bind_csv}
\title{Read in csv files with the same column structure and row-bind}
\usage{
bind_csv(path, pattern = NULL, save_csv = FALSE)
}
\arguments{
\item{path}{String containing the file path to the csv files.}

\item{pattern}{String containing pattern to match in the file name. Optional
and defaults to \code{NULL} (all csv files in path will be read in).}

\item{save_csv}{Logical, defaults to \code{FALSE}. If set to \code{TRUE}, a csv file
containing the combined data frame is saved in the same folder as the
provided path.}
}
\value{
When \code{save_csv} is set to \code{FALSE}, a data frame is returned. When set to
\code{TRUE}, a csv file containing the combined data frame is saved in the same
folder as the provided path.
}
\description{
csv files with a matched pattern in the file name in the given file
path are read in, and a row-bind operation is run to combine them. Pattern
matching is optional, and the user can choose whether to return a combined
data frame or save the output with the same name in the directory.
}
\section{How to Run}{
\preformatted{# Read files in and save output to path
bind_csv(
  path = "data",
  pattern = "2021-04-05",
  save_csv = TRUE
)

# Read files in and return data frame
bind_csv(
  path = "data"
)
}
}

