% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{camel_clean}
\alias{camel_clean}
\title{Convert "CamelCase" to "Camel Case"}
\usage{
camel_clean(string)
}
\arguments{
\item{string}{A string vector in 'CamelCase' format to format}
}
\description{
Convert a text string from the format "CamelCase" to "Camel Case".
This is used for converting variable names such as
"LevelDesignation" to "Level Designation" for the purpose
of prettifying plot labels.
}
\examples{
camel_clean("NoteHowTheStringIsFormatted")

}
\seealso{
Other Support: 
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{rgb2hex}()}
}
\concept{Support}
