% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_person_query}
\alias{check_person_query}
\title{Check a Person Query to ensure that it is suitable for analysis}
\usage{
check_person_query(data, return)
}
\arguments{
\item{data}{A person-level query in the form of a data frame. This includes:
\itemize{
\item Standard Person Query
\item Ways of Working Assessment Query
\item Hourly Collaboration Query
}

All person-level query have a \code{PersonId} column and a \code{Date} column.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"message"} (default)
\item \code{"text"}
}

See \code{Value} for more information.}
}
\description{
Prints diagnostic data about the data query to the R console, with information
such as date range, number of employees, HR attributes identified, etc.
}
\details{
Used as part of \code{check_query()}.
}
