% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{check_pkg_installed}
\alias{check_pkg_installed}
\title{Check whether package is installed and return an error message}
\usage{
check_pkg_installed(pkgname)
}
\arguments{
\item{pkgname}{String containing the name of the package to check whether is
installed.}
}
\description{
Checks whether a package is installed in the user's machine
based on a search on the package name string. If the package is not
installed, an error message is returned.
}
