% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_query_validation}
\alias{check_query_validation}
\title{Perform a query check for the validation report}
\usage{
check_query_validation(data, return)
}
\arguments{
\item{data}{A person-level query in the form of a data frame, including the standard person query,
Ways of Working assessment query, and the hourly collaboration query.}

\item{return}{Character string to specify whether to return a console message ("message"),
a character string ("text"), or a data frame. Defaults to "message".}
}
\description{
Prints diagnostic data about the data query to the R console, with information
such as date range, number of employees, HR attributes identified, etc.
Optimised for the \code{validation_report()}
}
\details{
Used as part of \code{check_query()}.
}
