% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaching_report.R
\name{coaching_report}
\alias{coaching_report}
\title{Generate a Coaching report in HTML}
\usage{
coaching_report(
  data,
  hrvar = "LevelDesignation",
  mingroup = 5,
  path = "coaching report",
  timestamp = TRUE
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "LevelDesignation"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file extension}.
For example, "collaboration report".}

\item{timestamp}{Logical vector specifying whether to include a timestamp in the file name.
Defaults to TRUE.}
}
\description{
The function generates an interactive HTML report using
Standard Person Query data as an input. The report contains a series
of summary analysis and visualisations relating to key \strong{coaching}
metrics in Workplace Analytics, specifically relating to the time
spent between managers and their direct reports.
}
\seealso{
Other Reports: 
\code{\link{IV_report}()},
\code{\link{capacity_report}()},
\code{\link{collaboration_report}()},
\code{\link{connectivity_report}()},
\code{\link{generate_report}()},
\code{\link{validation_report}()}
}
\concept{Reports}
