% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_fizz.R
\name{collaboration_fizz}
\alias{collaboration_fizz}
\alias{collab_fizz}
\title{Distribution of Collaboration Hours (Fizzy Drink plot)}
\usage{
collaboration_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")

collab_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\description{
Analyze weekly collaboration hours distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\details{
Uses the metric \code{Collaboration_hours}.
}
\examples{
collaboration_fizz(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Collaboration: 
\code{\link{collaboration_area}()},
\code{\link{collaboration_dist}()},
\code{\link{collaboration_line}()},
\code{\link{collaboration_rank}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()}
}
\concept{Collaboration}
