% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_line.R
\name{collaboration_line}
\alias{collaboration_line}
\title{Collaboration Time Trend - Line Chart}
\usage{
collaboration_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of collaboration time, visualised as line charts.
By default returns a line chart for collaboration hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{Collaboration_hours}.
}
\examples{

## Return a line plot
collaboration_line(sq_data, hrvar = "LevelDesignation")


## Return a table
collaboration_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Collaboration: 
\code{\link{collaboration_area}()},
\code{\link{collaboration_dist}()},
\code{\link{collaboration_fizz}()},
\code{\link{collaboration_rank}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()},
\code{\link{meeting_trend}()}
}
\concept{Collaboration}
