% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Box Plot for any metric}
\usage{
create_boxplot(
  data,
  metric,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A box plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Analyzes a selected metric and returns a a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\details{
This is a general purpose function that powers all the functions
in the package that produce box plots.
}
\examples{
# Create a fizzy plot for Work Week Span by Level Designation
create_boxplot(sq_data, metric = "Workweek_span", hrvar = "LevelDesignation", return = "plot")

# Create a summary statistics table for Work Week Span by Organization
create_boxplot(sq_data, metric = "Workweek_span", hrvar = "Organization", return = "table")

# Create a fizzy plot for Collaboration Hours by Level Designation
create_boxplot(sq_data, metric = "Collaboration_hours", hrvar = "LevelDesignation", return = "plot")
}
\seealso{
Other General: 
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_bubble}()},
\code{\link{create_scatter}()},
\code{\link{cut_hour}()},
\code{\link{extract_hr}()},
\code{\link{hrvar_count}()},
\code{\link{identify_outlier}()}
}
\concept{General}
