% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inc.R
\name{create_inc}
\alias{create_inc}
\title{Create an incidence table reflecting proportion of population scoring above
or below a threshold for a metric}
\usage{
create_inc(
  data,
  metric,
  hrvar,
  mingroup = 5,
  threshold,
  position,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{Character vector of length 2 containing the name of the HR
Variable by which to split metrics.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}

\item{threshold}{Numeric value specifying the threshold.}

\item{position}{String containing the below valid values:
\itemize{
\item \code{"above"}: show incidence of those equal to or above the threshold
\item \code{"below"}: show incidence of those equal to or below the threshold
}}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\description{
An incidence table is generated, with each value in the table reflecting the
proportion of the population that is above or below a threshold for a
specified metric. Two values are required for \code{hrvar} in order to generate
this incidence table.
}
\examples{
create_inc(
  data = sq_data,
  metric = "Collaboration_hours",
  hrvar = c("Organization", "LevelDesignation"),
  threshold = 20,
  position = "above"
)

}
