% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IV.R
\name{create_IV}
\alias{create_IV}
\title{Calculate Information Value for a selected outcome variable}
\usage{
create_IV(
  data,
  predictors = NULL,
  outcome,
  bins = 5,
  siglevel = 0.05,
  return = "plot"
)
}
\arguments{
\item{data}{A Person Query dataset in the form of a data frame.}

\item{predictors}{A character vector specifying the columns to be used as predictors.
Defaults to NULL, where all numeric vectors in the data will be used as predictors.}

\item{outcome}{A string specifying a binary variable, i.e. can only contain
the values 1 or 0.}

\item{bins}{Number of bins to use in \code{Information::create_infotables()}, defaults to 5.}

\item{siglevel}{Significance level to use in comparing populations for the outcomes,
defaults to 0.05}

\item{return}{String specifying what output to return.
\itemize{
\item \code{"plot"}: returns a bar plot summarising the information value with a maximum of 12 variables.
\item \code{"summary"} returns a summary table
\item \code{"list"} returns a list of outputs for all the input variables
\item \code{"plot-WOE"} commpares distribution for top predictors
\item \code{"IV"} returns the original Information object returned by \code{Information}.
}}
}
\description{
Specify an outcome variable and return IV outputs.
All numeric variables in the dataset are used as predictor variables.
}
\examples{
\dontrun{
library(dplyr)

# Return a summary table of IV
sq_data \%>\%
  mutate(X = ifelse(Email_hours > 6, 1, 0)) \%>\%
  create_IV(outcome = "X", return = "summary")

# Return plot
sq_data \%>\%
  mutate(X = ifelse(Collaboration_hours > 2, 1, 0)) \%>\%
  create_IV(outcome = "X",
            predictors = c("Email_hours", "Meeting_hours"),
            return = "plot")
}
}
\seealso{
Other Information Value: 
\code{\link{IV_by_period}()},
\code{\link{IV_report}()}
}
\concept{Information Value}
