% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_asis.R
\name{create_line_asis}
\alias{create_line_asis}
\title{Create a line chart without aggregation for any metric}
\usage{
create_line_asis(
  data,
  date_var = "Date",
  metric,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ylab = date_var,
  xlab = metric,
  line_colour = rgb2hex(0, 120, 212)
)
}
\arguments{
\item{data}{Plotting data as a data frame.}

\item{date_var}{String containing name of variable for the horizontal axis.}

\item{metric}{String containing name of variable representing the line.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{caption}{Caption of the plot.}

\item{ylab}{Y-axis label for the plot (group axis)}

\item{xlab}{X-axis label of the plot (bar axis).}

\item{line_colour}{String to specify colour to use for the line.
Hex codes are accepted. You can also supply
RGB values via \code{rgb2hex()}.}
}
\description{
This function creates a line chart directly from the aggregated / summarised data.
Unlike \code{create_line()} which performs a person-level aggregation, there is no
calculation for \code{create_line_asis()} and the values are rendered as they are passed
into the function. The only requirement is that a \code{date_var} is provided for the x-axis.
}
\seealso{
Other Flexible: 
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
