% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_period_scatter.R
\name{create_period_scatter}
\alias{create_period_scatter}
\title{Period comparison scatter plot for any two metrics}
\usage{
create_period_scatter(
  data,
  hrvar = "Organization",
  metric_x = "Multitasking_meeting_hours",
  metric_y = "Meeting_hours",
  before_start = min(as.Date(data$Date, "\%m/\%d/\%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$Date, "\%m/\%d/\%Y")),
  before_label = "Period 1",
  after_label = "Period 2",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector,
defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{metric_x}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{metric_y}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{before_start}{Start date of "before" time period in YYYY-MM-DD}

\item{before_end}{End date of "before" time period in YYYY-MM-DD}

\item{after_start}{Start date of "after" time period in YYYY-MM-DD}

\item{after_end}{End date of "after" time period in YYYY-MM-DD}

\item{before_label}{String to specify a label for the "before" period. Defaults to "Period 1".}

\item{after_label}{String to specify a label for the "after" period. Defaults to "Period 2".}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Returns two side-by-side scatter plots representing two selected metrics,
using colour to map an HR attribute and size to represent number of employees.
Returns a faceted scatter plot by default, with additional options
to return a summary table.
}
\details{
This is a general purpose function that powers all the functions
in the package that produce faceted scatter plots.
}
\examples{
\dontrun{

create_period_scatter(sq_data,
hrvar = "LevelDesignation",
before_start = "2019-11-03",
before_end = "2019-12-31",
after_start = "2020-01-01",
after_end = "2020-01-26")


create_period_scatter(sq_data, before_end = "2019-12-31", return = "table")
}

}
\seealso{
Other Flexible: 
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_rank}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
