% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank.R
\name{create_rank}
\alias{create_rank}
\title{Rank all groups across HR attributes on a selected Workplace Analytics metric}
\usage{
create_rank(
  data,
  metric,
  hrvar = extract_hr(data),
  mingroup = 5,
  return = "table"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{A character vector specifying what to return.
Valid values include "table" (default). Features are being considered for alternative return options but are currently
unavailable.}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
This function scans a standard Person query output for groups with high levels of a given Workplace Analytics Metric.
Returns a table with all groups (across multiple HR attributes) ranked by the specified metric.
}
\seealso{
Other Flexible: 
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
