% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank.R
\name{create_rank_combine}
\alias{create_rank_combine}
\title{Create combination pairs of HR variables and run 'create_rank()'}
\usage{
create_rank_combine(data, hrvar, metric, mingroup)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{String containing the name of the HR Variable by which to split
metrics. Defaults to \code{"Organization"}. To run the analysis on the total
instead of splitting by an HR attribute, supply \code{NULL} (without quotes).}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}
}
\description{
Create combination pairs of HR variables and run 'create_rank()'
}
