% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracking.R
\name{create_tracking}
\alias{create_tracking}
\title{Create a line chart that tracks metrics over time with rolling
averages}
\usage{
create_tracking(
  data,
  metric,
  plot_title = us_to_space(metric),
  plot_subtitle = "Measure over time",
  percent = FALSE
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"
percentage signs. Defaults to \code{FALSE}.}

\item{plot_title}{An option to override plot title.}

\item{plot_subtitle}{An option to override plot subtitle.}

\item{percent}{Logical value to determine whether to show labels as
percentage signs. Defaults to \code{FALSE}.}
}
\description{
Create a two-series line chart that visualizes a set of metric
over time for the selected population, with one of the series being a
four-week rolling average.
}
