% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trend.R
\name{create_trend}
\alias{create_trend}
\title{Heat mapped horizontal bar plot over time for any metric}
\usage{
create_trend(
  data,
  metric,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of a selected Workplace Analytics metric.
By default returns a week by week heatmap bar plot, highlighting the points in time with most activity.
Additional options available to return a summary table.
}
\examples{
create_trend(sq_data, metric = "Collaboration_hours", hrvar = "LevelDesignation")

}
\seealso{
Other Flexible: 
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_stacked}()}
}
\concept{Flexible}
