% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_data.R
\docType{data}
\name{em_data}
\alias{em_data}
\title{Sample Hourly Collaboration data with Email and IMs}
\format{
A data frame with 41567 rows and 56 variables:
\describe{
\item{PersonId}{ }
\item{Date}{ }
\item{IMs_sent_23_24}{ }
\item{IMs_sent_22_23}{ }
\item{IMs_sent_21_22}{ }
\item{IMs_sent_20_21}{ }
\item{IMs_sent_19_20}{ }
\item{IMs_sent_18_19}{ }
\item{IMs_sent_17_18}{ }
\item{IMs_sent_16_17}{ }
\item{IMs_sent_15_16}{ }
\item{IMs_sent_14_15}{ }
\item{IMs_sent_13_14}{ }
\item{IMs_sent_12_13}{ }
\item{IMs_sent_11_12}{ }
\item{IMs_sent_10_11}{ }
\item{IMs_sent_09_10}{ }
\item{IMs_sent_08_09}{ }
\item{IMs_sent_07_08}{ }
\item{IMs_sent_06_07}{ }
\item{IMs_sent_05_06}{ }
\item{IMs_sent_04_05}{ }
\item{IMs_sent_03_04}{ }
\item{IMs_sent_02_03}{ }
\item{IMs_sent_01_02}{ }
\item{IMs_sent_00_01}{ }
\item{Emails_sent_23_24}{ }
\item{Emails_sent_22_23}{ }
\item{Emails_sent_21_22}{ }
\item{Emails_sent_20_21}{ }
\item{Emails_sent_19_20}{ }
\item{Emails_sent_18_19}{ }
\item{Emails_sent_17_18}{ }
\item{Emails_sent_16_17}{ }
\item{Emails_sent_15_16}{ }
\item{Emails_sent_14_15}{ }
\item{Emails_sent_13_14}{ }
\item{Emails_sent_12_13}{ }
\item{Emails_sent_11_12}{ }
\item{Emails_sent_10_11}{ }
\item{Emails_sent_09_10}{ }
\item{Emails_sent_08_09}{ }
\item{Emails_sent_07_08}{ }
\item{Emails_sent_06_07}{ }
\item{Emails_sent_05_06}{ }
\item{Emails_sent_04_05}{ }
\item{Emails_sent_03_04}{ }
\item{Emails_sent_02_03}{ }
\item{Emails_sent_01_02}{ }
\item{Emails_sent_00_01}{ }
\item{Domain}{ }
\item{LevelDesignation}{ }
\item{Layer}{ }
\item{Organization}{ }
\item{IsInternal}{ }
\item{IsActive}{ }

...
}
}
\source{
\url{https://workplaceanalytics-demo.office.com/en-us/Home}
}
\usage{
em_data
}
\value{
data frame.
}
\description{
A sample dataset generated from an Hourly Collaboration query from the
Workplace Analytics demo tenant. The data is grouped by week and contains
columns for both IMs sent and Emails sent.
}
\seealso{
Other Data: 
\code{\link{dv_data}},
\code{\link{g2g_data}},
\code{\link{mt_data}},
\code{\link{p2p_data_sim}()},
\code{\link{sq_data}},
\code{\link{utc_data}}
}
\concept{Data}
\keyword{datasets}
