% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_fizz.R
\name{email_fizz}
\alias{email_fizz}
\title{Distribution of Email Hours (Fizzy Drink plot)}
\usage{
email_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A jittered scatter plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Analyze weekly email hours distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\examples{

# Return plot
email_fizz(sq_data, hrvar = "Organization", return = "plot")

# Return summary table
email_fizz(sq_data, hrvar = "Organization", return = "table")

}
\seealso{
Other Emails: 
\code{\link{email_dist}()},
\code{\link{email_line}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()},
\code{\link{email_trend}()}
}
\concept{Emails}
