% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export {wpa} outputs to CSV, clipboard, or save as images}
\usage{
export(
  x,
  method = "clipboard",
  path = "wpa export",
  timestamp = TRUE,
  width = 12,
  height = 9
)
}
\arguments{
\item{x}{Data frame or ggplot object to be passed through.}

\item{method}{Character string specifying the method of export.
Valid inputs include "clipboard" (default), "csv", "png", "svg", "jpeg", and "pdf".}

\item{path}{If exporting a file, enter the path and the desired file name, \emph{excluding the file extension}.
For example, "Analysis/SQ Overview".}

\item{timestamp}{Logical vector specifying whether to include a timestamp in the file name. Defaults to TRUE.}

\item{width}{Width of the plot}

\item{height}{Height of the plot}
}
\description{
A general use function to export {wpa} outputs to CSV,
clipboard, or save as images. By default, \code{export()} copies
a data frame to the clipboard. If the input is a ggplot object,
the default behaviour is to export a PNG.
}
