% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hr.R
\name{extract_hr}
\alias{extract_hr}
\title{Extract HR attribute variables}
\usage{
extract_hr(data, max_unique = 50, exclude_constants = TRUE, return = "names")
}
\arguments{
\item{data}{A data frame to be passed through.}

\item{max_unique}{A numeric value representing the maximum
number of unique values to accept for an HR attribute. Defaults to 50.}

\item{exclude_constants}{Logical value to specify whether single-value HR
attributes are to be excluded. Defaults to \code{TRUE}.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"names"}
\item \code{"vars"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"names"}: character vector identifying all the names of HR variables
present in the data.
\item \code{"vars"}: data frame containing all the columns of HR variables present
in the data.
}
}
\description{
This function uses a combination of variable class,
number of unique values, and regular expression matching
to extract HR / organisational attributes from a data frame.
}
\examples{
sq_data \%>\% extract_hr(return = "names")

sq_data \%>\% extract_hr(return = "vars")

}
\seealso{
Other General: 
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_scatter}()},
\code{\link{cut_hour}()},
\code{\link{hrvar_count}()},
\code{\link{identify_outlier}()}
}
\concept{General}
