% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_em_ratio.R
\name{flag_em_ratio}
\alias{flag_em_ratio}
\title{Flag unusual high Email hours to emails sent ratio}
\usage{
flag_em_ratio(data, threshold = 1)
}
\arguments{
\item{data}{A data frame containing a Person Query.}

\item{threshold}{Numeric value specifying the threshold for flagging. Defaults to 1.}
}
\description{
This function flags persons who have an unusual ratio
of email hours to emails sent.
Returns a character string by default.
}
\examples{
flag_em_ratio(sq_data)

}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
