% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_index.R
\name{flex_index}
\alias{flex_index}
\title{Compute a Flexibility Index based on the Hourly Collaboration Query}
\usage{
flex_index(
  data,
  hrvar = NULL,
  signals = c("email", "IM"),
  active_threshold = 0,
  start_hour = "0900",
  end_hour = "1700",
  return = "plot",
  plot_method = "common"
)
}
\arguments{
\item{data}{Hourly Collaboration query to be passed through as data frame.}

\item{hrvar}{A string specifying the HR attribute to cut the data by.
Defaults to NULL. This only affects the function when "table" is returned.}

\item{signals}{Character vector to specify which collaboration metrics to use:
You may use "email" for emails only, "IM" for Teams messages only,
or a combination of the two \code{c("email", "IM")} (default).}

\item{active_threshold}{A numeric value specifying the minimum number of signals to be greater than in
order to qualify as \emph{active}. Defaults to 0.}

\item{start_hour}{A character vector specifying starting hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying end hours,
e.g. "1700"}

\item{return}{Character vector to specify what to return.
Valid options include "plot" (default), "data", and "table".
When returning a plot, a random of ten working patterns are displayed, with diagnostic data and the Flexibility
Index shown on the plot.}

\item{plot_method}{Character string for determining which plot to return.
Options include "sample", "common", and "time". "sample"
plots a sample of ten working patterns; "common" plots the ten most common
working patterns; "time" plots the Flexibility Index for the group over time.}
}
\description{
Pass an Hourly Collaboration query and compute a Flexibility Index for a group
}
\examples{
# Examples of how to test the plotting options individually
# Sample of 10 work patterns
em_data \%>\%
  flex_index(return = "plot", plot_method = "sample")

# 10 most common work patterns
em_data \%>\%
  flex_index(return = "plot", plot_method = "common")

# Plot Flexibility Index over time
em_data \%>\%
  flex_index(return = "plot", plot_method = "time")

# Return a summary table with the computed Flexibility Index
em_data \%>\%
  flex_index(hrvar = "Organization", return = "table")

\dontrun{
# Return the raw data with the computed Flexibility Index
em_data \%>\%
  flex_index(return = "data")


}
}
\seealso{
Other Work Patterns: 
\code{\link{personas_hclust}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify_pav}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
