% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2g_data.R
\docType{data}
\name{g2g_data}
\alias{g2g_data}
\title{Sample Group-to-Group dataset}
\format{
A data frame with 3517 rows and 7 variables:
\describe{
\item{TimeInvestors_Organization}{ }
\item{Collaborators_Organization}{ }
\item{Date}{ }
\item{Meetings}{ }
\item{Meeting_hours}{ }
\item{Email_hours}{ }
\item{Collaboration_hours}{ }

...
}
}
\source{
\url{https://workplaceanalytics-demo.office.com/en-us/Home}
}
\usage{
g2g_data
}
\description{
A demo dataset generated from a Group-to-Group Query from Workplace Analytics.
}
\keyword{datasets}
