% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2g_network.R
\name{g2g_network}
\alias{g2g_network}
\title{Create a network plot with the G2G query}
\usage{
g2g_network(
  data,
  time_investor,
  collaborator,
  metric,
  exc_threshold = 0.1,
  subtitle = "Collaboration Across Organizations",
  return = "plot",
  ...
)
}
\arguments{
\item{data}{Data frame containing a G2G query.}

\item{time_investor}{String containing the variable name for the Time Investor column.}

\item{collaborator}{String containing the variable name for the Collaborator column.}

\item{metric}{String containing the variable name for metric.}

\item{exc_threshold}{Exclusion threshold to apply.}

\item{subtitle}{String to override default plot subtitle.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{...}{Additional arguments to pass to \code{GGally::ggnet2()}}
}
\description{
Pass a data frame containing a G2G query and return a network
plot.
Automatically handles "Collaborators_within_group" and "Other_collaborators" within query data.
}
\examples{
\dontrun{
g2g_data \%>\%
  g2g_network(time_investor = "TimeInvestors_Organization",
              collaborator = "Collaborators_Organization",
              metric = "Collaboration_hours")
}

}
