% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_chunks}
\alias{generate_chunks}
\title{Generate chunk strings}
\usage{
generate_chunks(level = 3, title, subheader = "", echo = FALSE, object)
}
\arguments{
\item{level}{Numeric value to specify the header level of the chunk.}

\item{title}{Character string to specify the title of the chunk.}

\item{subheader}{Character string to specify the subheader of the chunk.}

\item{echo}{Logical value to specify whether to display code.}

\item{object}{Character string to specify name of the object to show.}
}
\description{
This is used as a supporting function for \code{generate_report()}
and not directly used. `generate_report()`` works by creating a
loop structure around generate_chunks(), and binds them together
to create a report.
}
\details{
\code{generate_chunks()} is primarily a wrapper around paste() functions,
to create a structured character vector that will form the individual
chunks. No plots 'exist' within the environment of \code{generate_chunks()}.
}
