% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_zoom.R
\name{hash_zoom}
\alias{hash_zoom}
\title{Reads in unhashed Zoom meeting participants file and merges data with
a hash file}
\usage{
hash_zoom(zoom_path, hash_path, match_only = FALSE)
}
\arguments{
\item{zoom_path}{string containing the path to the unhashed Zoom participants
.csv file.}

\item{match_only}{logical. Determines whether to include only rows where
there is a corresponding match on the hash file. Defaults to \code{FALSE}.}

\item{hash_file}{string containing the path to the .csv file containing the
hash IDs. The file should contain only two columns with the following
headers:
\itemize{
\item \code{PersonID}: column containing email addresses that map to the Zoom
\code{User_Name} columns.
\item \code{HashID}: column containing HashIDs which will be uploaded as
organizational attributes to Workplace Analytics.
}}
}
\description{
Reads in an unhashed Zoom meeting participants file from .csv and replaces
usernames with a \code{HashID} that is supplied in a separate look up file. This
is to be used \emph{after} \code{bind_csv()} which will combine all the separate Zoom
participants file into a single file.
}
