% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_colours.R
\name{heat_colours}
\alias{heat_colours}
\title{Generate a vector of \code{n} contiguous colours, as a red-yellow-green palette.}
\usage{
heat_colours(n, alpha, rev = FALSE)
}
\arguments{
\item{n}{the number of colors (≥ 1) to be in the palette.}

\item{alpha}{an alpha-transparency level in the range \link{0,1} (0 means transparent and 1 means opaque)}

\item{rev}{logical indicating whether the ordering of the colors should be reversed.}
}
\description{
Generate a vector of \code{n} contiguous colours, as a red-yellow-green palette.
}
\examples{
barplot(rep(10, 50), col = heat_colours(n = 50), border = NA)

barplot(rep(10, 50), col = heat_colours(n = 50, alpha = 0.5, rev = TRUE), border = NA)

}
