% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_trend.R
\name{hr_trend}
\alias{hr_trend}
\title{Employee count over time}
\usage{
hr_trend(data, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Returns a line chart showing the change in
employee count over time. Part of a data validation process to check
for unusual license growth / declines over time.
}
\examples{
# Return a plot
hr_trend(dv_data)

# Return a table
hr_trend(dv_data, return = "table")

}
