% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_shifts.R
\name{identify_shifts}
\alias{identify_shifts}
\title{Identify shifts based on outlook time settings for work day start and end time}
\usage{
identify_shifts(data, return = "plot")
}
\arguments{
\item{data}{A data frame containing data from the Hourly Collaboration query.}

\item{return}{Character vector to specify what to return.
"table" (default) returns a summary table of the counts of shifts.
"data" returns the original input data frame with the \code{Shifts} column appended.}
}
\description{
This function uses outlook calendar settings for start and end time of work day to identify
work shifts. The relevant variables are \code{WorkingStartTimeSetInOutlook} and
\code{WorkingEndTimeSetInOutlook}.
}
\examples{
dv_data \%>\% identify_shifts()

}
