% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tenure.R
\name{identify_tenure}
\alias{identify_tenure}
\title{Tenure calculation based on different input dates, returns data summary table or histogram}
\usage{
identify_tenure(
  data,
  end_date = "Date",
  beg_date = "HireDate",
  maxten = 40,
  return = "message"
)
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{end_date}{A string specifying the name of the date variable representing the latest date. Defaults to "Date".}

\item{beg_date}{A string specifying the name of the date variable representing the hire date. Defaults to "HireDate".}

\item{maxten}{A numeric value representing the maximum tenure.
If the tenure exceeds this threshold, it would be accounted for in the flag message.}

\item{return}{String to specify what to return.
Defaults to "message".
Other valid values include "text", "plot", "data_cleaned", "data_dirty", and "data".
For "data", a data frame with the \code{PersonId} and a calculated variable called \code{TenureYear} is returned.}
}
\description{
This function calculates employee tenure based on different input dates.
\code{identify_tenure} uses the latest Date available if user selects "Date",
but also have flexibility to select a specific date, e.g. "1/1/2020".
}
\examples{
\dontrun{
# Add HireDate to sq_data - method #1
sq_data2 <- sq_data \%>\% mutate(HireDate = as.Date("1/1/2015", format = "\%m/\%d/\%Y" ))
identify_tenure(sq_data2)

# Add HireDate to sq_data - method #2
sq_data$HireDate <-
  rep(sample(seq(as.Date('1975/01/01'),
                 as.Date('2019/11/01'), by="day"), 15119),2)

identify_tenure(sq_data)
}

}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
