% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zoom.R
\name{import_zoom}
\alias{import_zoom}
\title{Import a Zoom Meeting Participants csv file}
\usage{
import_zoom(path, date_format = "\%m/\%d/\%Y \%H:\%M:\%S")
}
\description{
A fast method to import a Zoom Meeting Participants csv file whilst
preserving UTF-8 encoding. A convenience wrapper around
\code{data.table::fread()}.
}
