% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_network_plot.R
\name{internal_network_plot}
\alias{internal_network_plot}
\title{Plot Internal Network Breadth and Size as a scatter plot}
\usage{
internal_network_plot(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  bubble_size = c(1, 8)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to
"Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the
following strings: - \code{"plot"} - \code{"table"}}

\item{bubble_size}{A numeric vector of length two to specify the size range
of the bubbles}
}
\description{
Plot the internal network metrics for a HR variable as a scatter plot, showing
Internal Network Breadth as the vertical axis and Internal Network Size as the
horizontal axis.
}
\examples{
# Return plot
sq_data \%>\% internal_network_plot(return = "plot")

}
\concept{Connectivity}
