% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_by_period.R
\name{IV_by_period}
\alias{IV_by_period}
\title{Identify the WPA metrics that have the biggest change between two periods.}
\usage{
IV_by_period(
  data,
  before_start = min(as.Date(data$Date, "\%m/\%d/\%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$Date, "\%m/\%d/\%Y")),
  mybins = 10,
  return = "table"
)
}
\arguments{
\item{data}{Person Query as a dataframe including date column named "Date"
This function assumes the data format is MM/DD/YYYY as is standard in a WpA query output.}

\item{before_start}{Start date of "before" time period in YYYY-MM-DD. Defaults to earliest date in dataset.}

\item{before_end}{End date of "before" time period in YYYY-MM-DD}

\item{after_start}{Start date of "after" time period in YYYY-MM-DD. Defaults to day after before_end.}

\item{after_end}{End date of "after" time period in YYYY-MM-DD. Defaults to latest date in dataset.}

\item{mybins}{Number of bins to cut the data into for Information Value analysis. Defaults to 10.}

\item{return}{Specify whether to return a summary table or detailed Excel and PDF files
Specify "table" or "detailed" for outputs. Defaults to table.}
}
\description{
This function uses the Information Value algorithm to predict which WPA metrics are most explained by the change in dates.
}
\seealso{
Other Flexible Input: 
\code{\link{period_change}()}

Other Information Value: 
\code{\link{IV_report}()},
\code{\link{create_IV}()}
}
\concept{Flexible Input}
\concept{Information Value}
