% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymetrics_scan.R
\name{keymetrics_scan}
\alias{keymetrics_scan}
\title{Run a summary of Key Metrics from the Standard Person Query data}
\usage{
keymetrics_scan(
  data,
  hrvar = "Organization",
  mingroup = 5,
  metrics = c("Workweek_span", "Collaboration_hours",
    "After_hours_collaboration_hours", "Meetings", "Meeting_hours",
    "After_hours_meeting_hours", "Low_quality_meeting_hours",
    "Meeting_hours_with_manager_1_on_1", "Meeting_hours_with_manager", "Emails_sent",
    "Email_hours", "After_hours_email_hours", "Generated_workload_email_hours",
    "Total_focus_hours", "Internal_network_size", "Networking_outside_organization",
    "External_network_size", "Networking_outside_company"),
  return = "plot",
  textsize = 2
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{String containing the name of the HR Variable by which to split
metrics. Defaults to \code{"Organization"}. To run the analysis on the total
instead of splitting by an HR attribute, supply \code{NULL} (without quotes).}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}

\item{metrics}{A character vector containing the variable names to calculate
averages of.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{textsize}{A numeric value specifying the text size to show in the plot.}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Returns a heatmapped table by default, with options to return a table.
}
\examples{
# Heatmap plot is returned by default
keymetrics_scan(sq_data)

# Return summary table
keymetrics_scan(sq_data, hrvar = "LevelDesignation", return = "table")

}
