% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin.R
\name{maxmin}
\alias{maxmin}
\title{Max-Min Scaling Function}
\usage{
maxmin(x)
}
\arguments{
\item{x}{Pass a vector or the required columns of a data frame through this argument.}
}
\description{
This function allows you to scale vectors or an entire data frame using the max-min scaling method
A numeric vector is always returned. Originally implemented in https://github.com/martinctc/surveytoolbox.
}
\details{
This is used within \code{keymetrics_scan()} to enable row-wise heatmapping.
}
\examples{
numbers <- c(15, 40, 10, 2)
maxmin(numbers)
}
\keyword{max-min}
