% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{md2html}
\alias{md2html}
\title{Display HTML fragment in RMarkdown chunk, from Markdown text}
\usage{
md2html(text)
}
\arguments{
\item{text}{Character vector containing Markdown text}
}
\description{
This is a wrapper around \code{markdown::markdownToHTML()}, where
the default behaviour is to produce a HTML fragment.
\code{htmltools::HTML()} is then used to evaluate the HTML code
within a RMarkdown chunk.
}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{is_date_format}()},
\code{\link{maxmin}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{plot_WOE}()},
\code{\link{read_preamble}()},
\code{\link{rgb2hex}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{tstamp}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}
}
\concept{Support}
