% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_dist.R
\name{meeting_dist}
\alias{meeting_dist}
\alias{meeting_distribution}
\title{Meeting Hours distribution}
\usage{
meeting_dist(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(5, 10, 15)
)

meeting_distribution(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(5, 10, 15)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{cut}{A numeric vector of length three to specify the breaks for the distribution,
e.g. c(10, 15, 20)}
}
\description{
Analyze Meeting Hours distribution.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
## Return a plot
meeting_dist(sq_data, hrvar = "Organization")

## Return a table
meeting_dist(sq_data, hrvar = "Organization", return = "table")

## Return result with a custom specified breaks
meeting_dist(sq_data, hrvar = "LevelDesignation", cut = c(4, 7, 9))


}
\seealso{
Other Meeting Culture: 
\code{\link{meeting_quality}()},
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()}
}
\concept{Meeting Culture}
