% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_fizz.R
\name{meeting_fizz}
\alias{meeting_fizz}
\title{Distribution of Meeting Hours (Fizzy Drink plot)}
\usage{
meeting_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze weekly meeting hours distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
meeting_fizz(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Meetings: 
\code{\link{meeting_line}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_summary}()}
}
\concept{Meetings}
