% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_line.R
\name{meeting_line}
\alias{meeting_line}
\title{Meeting Time Trend - Line Chart}
\usage{
meeting_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A faceted line plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Provides a week by week view of meeting time, visualised as line charts.
By default returns a line chart for meeting hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\examples{
# Return a line plot
meeting_line(sq_data, hrvar = "LevelDesignation")

# Return summary table
meeting_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Meetings: 
\code{\link{meeting_fizz}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_summary}()},
\code{\link{meeting_trend}()},
\code{\link{meetingtype_summary}()}
}
\concept{Meetings}
