% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_rank.R
\name{meeting_rank}
\alias{meeting_rank}
\title{Meeting Hours Ranking}
\usage{
meeting_rank(data, hrvar = extract_hr(data), mingroup = 5, return = "table")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"table"} (default)
\item \code{"plot"}
}

See \code{Value} for more information.}
}
\value{
When 'table' is passed in \code{return}, a summary table is returned as a data frame.
}
\description{
This function scans a standard query output for groups with high levels of 'Weekly Email Collaboration'.
Returns a table with a all of groups (across multiple HR attributes) ranked by hours of digital collaboration.
}
\details{
Uses the metric \code{Meeting_hours}.
See \code{create_rank()} for applying the same analysis to a different metric.
}
\seealso{
Other Meetings: 
\code{\link{meeting_fizz}()},
\code{\link{meeting_line}()},
\code{\link{meeting_summary}()},
\code{\link{meeting_trend}()},
\code{\link{meetingtype_summary}()}
}
\concept{Meetings}
