% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_skim.R
\name{meeting_skim}
\alias{meeting_skim}
\title{Produce a skim summary of meeting hours}
\usage{
meeting_skim(data, return = "message")
}
\arguments{
\item{data}{A standard person query data in the form of a data frame.}

\item{return}{A character string to specify whether to return a message
or a text string. Defaults to "message". Valid options include "text" or
"table".}
}
\description{
This function returns a skim summary in the console
when provided a standard query in the input.
}
\examples{
meeting_skim(sq_data)

}
