% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_tm_report.R
\name{meeting_tm_report}
\alias{meeting_tm_report}
\title{Generate a Meeting Text Mining report in HTML}
\usage{
meeting_tm_report(
  data,
  path = "meeting text mining report",
  stopwords = NULL,
  timestamp = TRUE,
  keep = 100,
  seed = 100
)
}
\arguments{
\item{data}{A Meeting Query dataset in the form of a data frame.}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file extension}.
For example, "meeting text mining report".}

\item{stopwords}{A single-column data frame labelled 'word' containing custom stopwords to remove.}

\item{timestamp}{Logical vector specifying whether to include a timestamp in the file name.
Defaults to TRUE.}

\item{keep}{A numeric vector specifying maximum number of words to keep.}

\item{seed}{A numeric vector to set seed for random generation.}
}
\description{
Create a text mining report in HTML based on Meeting Subject Lines
}
\seealso{
Other Text-mining: 
\code{\link{tm_clean}()},
\code{\link{tm_cooc}()},
\code{\link{tm_freq}()},
\code{\link{tm_wordcloud}()}
}
\concept{Text-mining}
