% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetingtype_dist_ca.R
\name{meetingtype_dist_ca}
\alias{meetingtype_dist_ca}
\title{Meeting Type Distribution (Collaboration Assessment Query)}
\usage{
meetingtype_dist_ca(data, hrvar = NULL, mingroup = 5, return = "plot")
}
\arguments{
\item{data}{Meeting Query data frame. Must contain the variables \code{Attendee} and \code{DurationHours}}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Calculate the hour distribution of internal meeting types,
using a Collaboration Assessment Query with core WpA variables as an input.
}
