% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrcoatt_dist.R
\name{mgrcoatt_dist}
\alias{mgrcoatt_dist}
\title{Manager meeting coattendance distribution}
\usage{
mgrcoatt_dist(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze degree of attendance between employes and their managers.
Returns a stacked bar plot of different buckets of coattendance.
Additional options available to return a table with distribution elements.
}
\examples{
workloads_dist(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Managerial Relations: 
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_fizz}()},
\code{\link{one2one_line}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}

Other Meeting Culture: 
\code{\link{meeting_dist}()},
\code{\link{meeting_quality}()},
\code{\link{mgrrel_matrix}()}
}
\concept{Managerial Relations}
\concept{Meeting Culture}
