% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrrel_matrix.R
\name{mgrrel_matrix}
\alias{mgrrel_matrix}
\title{Manager Relationship 2x2 Matrix}
\usage{
mgrrel_matrix(
  data,
  hrvar = NULL,
  return = "plot",
  plot_colors = c("#FE7F4F", "#C3E5FF", "#FFD4C4", "#FEAA8A"),
  threshold = 15
)
}
\arguments{
\item{data}{Standard Query data to pass through. Accepts a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector,
e.g. "Organization". Defaults to NULL.}

\item{return}{A character vector specifying whether to return a matrix plot or a table.
Defaults to the 2 by 2 matrix. Valid values are "plot", "table", and "chartdata".}

\item{plot_colors}{Pass a character vector of length 4 containing HEX codes to specify colors to use in plotting.}

\item{threshold}{Specify a numeric value to determine threshold (in minutes) for 1:1 manager hours.
Defaults to 15.}
}
\description{
Generate the Manager-Relationship 2x2 matrix, returning a ggplot object by default.
Additional options available to return a "wide" or "long" summary table.
}
\examples{
mgrrel_matrix(sq_data)

}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_line}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}

Other Meeting Culture: 
\code{\link{meeting_dist}()},
\code{\link{meeting_quality}()},
\code{\link{mgrcoatt_dist}()}
}
\concept{Managerial Relations}
\concept{Meeting Culture}
