% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_data.R
\docType{data}
\name{mt_data}
\alias{mt_data}
\title{Sample Meeting Query dataset}
\format{
A data frame with 2000 rows and 30 variables:
\describe{
\item{MeetingId}{ }
\item{StartDate}{ }
\item{StartTimeUTC}{ }
\item{EndDate}{ }
\item{EndTimeUTC}{ }
\item{Attendee_meeting_hours}{ }
\item{Attendees}{ }
\item{Organizer_Domain}{ }
\item{Organizer_FunctionType}{ }
\item{Organizer_LevelDesignation}{ }
\item{Organizer_Layer}{ }
\item{Organizer_Region}{ }
\item{Organizer_Organization}{ }
\item{Organizer_zId}{ }
\item{Organizer_attainment}{ }
\item{Organizer_TimeZone}{ }
\item{Organizer_HourlyRate}{ }
\item{Organizer_IsInternal}{ }
\item{Organizer_PersonId}{ }
\item{IsCancelled}{ }
\item{DurationHours}{ }
\item{IsRecurring}{ }
\item{Subject}{ }
\item{TotalAccept}{ }
\item{TotalNoResponse}{ }
\item{TotalDecline}{ }
\item{TotalNoEmailsDuringMeeting}{ }
\item{TotalNoDoubleBooked}{ }
\item{TotalNoAttendees}{ }
\item{MeetingResources}{ }

...
}
}
\source{
\url{https://workplaceanalytics-demo.office.com/en-us/Home}
}
\usage{
mt_data
}
\description{
A dataset generated from a Meeting Query from WpA.
}
\keyword{datasets}
