% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_leiden.R
\name{network_leiden}
\alias{network_leiden}
\title{Implement the Leiden community detection on a Person to Person network query}
\usage{
network_leiden(data, hrvar, res = 0.5, return)
}
\arguments{
\item{data}{Data frame containing a Person to Person query.}

\item{hrvar}{String containing the HR attribute to be matched in the dataset.}

\item{res}{Resolution parameter to be passed to \code{leiden::leiden()}. Defaults to 0.5.}

\item{return}{String specifying what output to return. Valid return options include:
\itemize{
\item 'plot-leiden': return a network plot coloured by leiden communities.
\item 'plot-hrvar': return a network plot coloured by HR attribute.
\item 'plot-sankey': return a sankey plot combining communities and HR attribute.
\item 'table': return a vertex summary table with counts in communities and HR attribute.
\item 'data': return a vertex data file that matches vertices with communities and HR attributes.
}}
}
\description{
Take a P2P network query and implement the Leiden community detection method. To run
this function, you will require all the pre-requisites of the \strong{leiden} package installed,
which includes Python and \strong{reticulate}.
}
