% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_louvain.R
\name{network_louvain}
\alias{network_louvain}
\title{Implement the Louvain community detection on a Person to Person network query}
\usage{
network_louvain(data, hrvar, return)
}
\arguments{
\item{data}{Data frame containing a Person to Person query.}

\item{hrvar}{String containing the HR attribute to be matched in the dataset.}

\item{return}{String specifying what output to return.}
}
\description{
Take a P2P network query and implement the Louvain community detection method. The
\strong{igraph} implementation of the Louvain method is used.
}
