% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_p2p_test.R
\name{network_p2p_test}
\alias{network_p2p_test}
\title{Create a network plot with the person-to-person query}
\usage{
network_p2p_test(
  data,
  hrvar,
  centrality = NULL,
  community = NULL,
  return,
  bg_fill = "#FFFFFF",
  font_col = "grey20",
  legend_pos = "bottom",
  palette = "rainbow",
  node_alpha = 0.7,
  edge_alpha = 1,
  res = 0.5,
  seed = 1,
  algorithm = "mds",
  size_threshold = 5000,
  weight = "StrongTieScore"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Analyse a person-to-person (P2P) network query, with multiple visualisation
and analysis output options. Pass a data frame containing a person-to-person
query and return a network visualization. Options are available for community
detection using either the Louvain or the Leiden algorithms.
}
\seealso{
Other Network: 
\code{\link{external_network_plot}()},
\code{\link{g2g_data}},
\code{\link{internal_network_plot}()},
\code{\link{network_describe}()},
\code{\link{network_g2g}()},
\code{\link{network_leiden}()},
\code{\link{network_louvain}()},
\code{\link{network_p2p}()},
\code{\link{network_summary}()},
\code{\link{p2p_data_sim}()}
}
\concept{Network}
