% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_dist.R
\name{one2one_dist}
\alias{one2one_dist}
\title{Manager 1:1 Time Trend distribution}
\usage{
one2one_dist(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze Manager 1:1 Time distribution.
Returns a stacked bar plot of different buckets of 1:1 time.
Additional options available to return a table with distribution elements.
}
\examples{
workloads_dist(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_line}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}
}
\concept{Managerial Relations}
