% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_freq.R
\name{one2one_freq}
\alias{one2one_freq}
\title{Frequency of Manager 1:1 Meetings as bar or 100\% stacked bar chart}
\usage{
one2one_freq(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  mode = "dist"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}}

\item{mode}{String specifying what method to use. This must be one of the following strings:
\itemize{
\item \code{"sum"}
\item \code{"dist"}
}}
}
\description{
This function calculates the average number of weeks (cadence) between of 1:1 meetings between an employee and their manager.
Returns a  distribution plot for typical cadence of 1:1 meetings.
Additional options available to return a bar plot, tables, or a dataframe with a Cadence of 1 on 1 meetings metric.
}
\examples{
# Return plot
one2one_freq(sq_data, hrvar = "Organization", return = "plot")

# Return summary table
one2one_freq(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_fizz}()},
\code{\link{one2one_line}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}
}
\concept{Managerial Relations}
