% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_line.R
\name{one2one_line}
\alias{one2one_line}
\title{Manager 1:1 Time Trend - Line Chart}
\usage{
one2one_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A faceted line plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Provides a week by week view of 1:1 time with managers, visualised as line charts.
By default returns a line chart for 1:1 meeting hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{Meeting_hours_with_manager_1_on_1}.
}
\examples{
# Return a line plot
one2one_line(sq_data, hrvar = "LevelDesignation")

# Return summary table
one2one_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_fizz}()},
\code{\link{one2one_freq}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}
}
\concept{Managerial Relations}
