% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_change.R
\name{period_change}
\alias{period_change}
\title{Plot the distribution of percentage change between periods
of a WpA metric by the number of employees.}
\usage{
period_change(
  data,
  compvar,
  before_start = min(as.Date(data$Date, "\%m/\%d/\%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$Date, "\%m/\%d/\%Y")),
  return = "count"
)
}
\arguments{
\item{data}{Person Query as a dataframe including date column named "Date"
This function assumes the data format is MM/DD/YYYY as is standard in a WpA query output.}

\item{compvar}{WpA comparison variable to compare person change before and after
For example, "Collaboration_hours"}

\item{before_start}{Start date of "before" time period in YYYY-MM-DD}

\item{before_end}{End date of "before" time period in YYYY-MM-DD}

\item{after_start}{Start date of "after" time period in YYYY-MM-DD}

\item{after_end}{End date of "after" time period in YYYY-MM-DD}

\item{return}{Character vector specifying whether to return plot as Count or Percentage of Employees
Valid inputs include "count" (default), "percentage", and "table".}
}
\description{
This function also presents the p-value for the null hypothesis
that the variable has not changed, using a Wilcox signed-rank test.
}
\examples{
# Run plot
period_change(sq_data, compvar = "Workweek_span", before_end = "2019-11-16")

\donttest{
# Run plot with more specific arguments
period_change(sq_data,
              compvar = "Workweek_span",
              before_start = "2019-11-03",
              before_end = "2019-11-16",
              after_start = "2019-12-03",
              after_end = "2019-12-16",
              return = "percentage")
}
}
\seealso{
Other Flexible Input: 
\code{\link{IV_by_period}()}
}
\concept{Flexible Input}
