% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personas_hclust.R
\name{personas_hclust}
\alias{personas_hclust}
\title{Create a hierarchical clustering of selected metrics}
\usage{
personas_hclust(data, metrics, k = 4, return = "table")
}
\arguments{
\item{data}{A data frame containing \code{PersonId} and selected metrics for clustering.}

\item{metrics}{Vector of metrics to use for clustering.}

\item{k}{Numeric vector to specify the \code{k} number of clusters to cut by.}

\item{return}{Character vector to specify what to return.
Valid options include "data", "table" and "hclust".
"hclust" returns the hierarchical model generated by the function.}
}
\description{
Apply hierarchical clustering to selected metrics
The hierarchical clustering uses cosine distance and the ward.D method
of agglomeration.
}
\examples{
\dontrun{
personas_hclust(sq_data, k = 4, return = "table")

personas_hclust(sq_data, k = 4, return = "data")

}


}
\seealso{
Other Work Patterns: 
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
