% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_hclust.R
\name{plot_signal_clust}
\alias{plot_signal_clust}
\title{Plot signal hour patterns by clusters}
\usage{
plot_signal_clust(data, group_label, type = "bar", sig_label = "Emails_sent")
}
\arguments{
\item{data}{Data frame containing the normalised email by hour
data and the cluster variable}

\item{group_label}{Character vector specifying the name of the variable
containing the clusters.}

\item{type}{Character vector to specify type of plot to return.
Accepted values are "bar" (default) and "area".}

\item{sig_label}{Character vector to select required columns,
e.g. "Emails_sent", "IMs_sent", "Signals_sent"}
}
\description{
Workhorse function feeding into \code{workpatterns_hclust()}.
This is a build on \code{plot_email_clust} to allow flexible inputs.
}
