% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IV.R
\name{plot_WOE}
\alias{plot_WOE}
\title{Plot WOE graphs with an IV object}
\usage{
plot_WOE(IV, predictor)
}
\arguments{
\item{IV}{IV object created with 'Information'.}

\item{predictor}{String with the name of the predictor variable.}
}
\value{
ggplot object. Bar plot with 'WOE' as the y-axis and bins of the predictor
variable as the horizontal axis.
}
\description{
Internal function within \code{create_IV()} that plots WOE graphs using an IV
object. Can also be used for plotting individual WOE graphs.
}
