% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_classify_bw.R
\name{plot_workpatterns_classify_bw}
\alias{plot_workpatterns_classify_bw}
\title{Plotting function for \code{workpatterns_classify_bw()}}
\usage{
plot_workpatterns_classify_bw(data)
}
\description{
Internal use only.
}
