% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{read_preamble}
\alias{read_preamble}
\title{Read preamble}
\usage{
read_preamble(path)
}
\arguments{
\item{path}{Text string containing the path for the appropriate Markdown file.}
}
\description{
Read in a preamble to be used within each individual reporting function.
Reads from the Markdown file installed with the package.
}
