% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert rgb to HEX code}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r, g, b}{Values that correspond to the three RGB parameters}
}
\description{
Convert rgb to HEX code
}
