% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_hclust.R
\name{run_hour_splits}
\alias{run_hour_splits}
\title{Run working hour splits for signals sent}
\usage{
run_hour_splits(
  data,
  start_hour = "0900",
  end_hour = "1700",
  group_label = "cluster"
)
}
\arguments{
\item{data}{See \code{workpatterns_hclust()}.}

\item{start_hour}{See \code{workpatterns_hclust()}.}

\item{end_hour}{See \code{workpatterns_hclust()}.}

\item{group_label}{See \code{workpatterns_hclust()}.}
}
\description{
Used internally within \code{workpatterns_hclust()}
}
