% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_hclust.R
\name{run_sum_hr}
\alias{run_sum_hr}
\title{Run summary statistics for hours of day}
\usage{
run_sum_hr(data, group_label = "cluster", sig_label = "Emails_sent")
}
\arguments{
\item{data}{See \code{workpatterns_hclust()}.}

\item{group_label}{A character vector for the column name containing the clusters
or the groups. Currently accepted values are "cluster" and "Personas".}

\item{sig_label}{See \code{workpatterns_hclust()}.}
}
\description{
Supporting function that takes the raw data from \code{workpatterns_hclust()}
as an input.
}
