% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_hm.R
\name{split_hm}
\alias{split_hm}
\title{Takes in a numeric value of 'hhmm' format and convert this into a 'hms' format}
\usage{
split_hm(x)
}
\arguments{
\item{x}{String in 'hhmm' format.}
}
\description{
Takes in a numeric value of 'hhmm' format and convert this into a 'hms' format
}
