% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_scan.R
\name{subject_scan}
\alias{subject_scan}
\title{Count top words in subject lines grouped by a custom attribute}
\usage{
subject_scan(data, hrvar, n = 10, return = "plot")
}
\arguments{
\item{data}{A Meeting Query dataset in the form of a data frame.}

\item{hrvar}{String containing the name of the HR Variable by which to split
metrics.}

\item{n}{Numeric value specifying the top number of words to show.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
\item \code{"data"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"plot"}: 'ggplot' object. A heatmapped grid.
\item \code{"table"}: data frame. A summary table for the metric.
\item \code{"data"}: data frame.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function generates a matrix of the top occurring words in meetings,
grouped by a specified attribute such as organisational attribute, day of the
week, or hours of the day.
}
\examples{
mt_data \%>\% subject_scan()

}
