% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals_col.R
\name{totals_col}
\alias{totals_col}
\title{Fabricate a 'Total' HR variable}
\usage{
totals_col(data, total_value = "Total")
}
\arguments{
\item{data}{data frame}

\item{total_value}{Character value defining the name and the value of the "Total" column.
Defaults to "Total". An error is returned if an existing variable has the same name as the
supplied value.}
}
\description{
Create a 'Total' column of character type comprising exactly of one unique value.
This is a convenience function for returning a no-HR attribute view when NULL is supplied to
the \code{hrvar} argument in functions.
}
\examples{
# Create a visual without HR attribute breaks
sq_data \%>\%
  totals_col() \%>\%
  collab_fizz(hrvar = "Total")

}
