% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstamp.R
\name{tstamp}
\alias{tstamp}
\title{Generate a time stamp}
\usage{
tstamp()
}
\description{
This function generates a time stamp of the format yymmdd_hhmmss.
This is a support function and is not intended for direct use.
}
