% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_to_space.R
\name{us_to_space}
\alias{us_to_space}
\title{Replace underscore with space}
\usage{
us_to_space(x)
}
\arguments{
\item{x}{String to replace all occurrences of \verb{_} with a single space}
}
\description{
Convenience function to convert underscores to space
}
\examples{
us_to_space("Meeting_hours_with_manager_1_on_1")

}
