% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_data.R
\docType{data}
\name{utc_data}
\alias{utc_data}
\title{UTC offset dataset based on Workplace Analytics time zones}
\format{
A data frame with 128 rows and 4 variables:
\describe{
\item{WpA_TimeZone}{ }
\item{UTC}
\item{UTC_offset}
\item{UTC_offset_raw}

...
}
}
\source{
\url{https://docs.microsoft.com/en-us/windows-hardware/manufacture/desktop/default-time-zones/}

\url{https://docs.microsoft.com/en-us/workplace-analytics/use/timezones-for-workplace-analytics}
}
\usage{
utc_data
}
\value{
data frame.
}
\description{
A dataset containing the UTC offset values based on time zones used in
Workplace Analytics.
}
\seealso{
Other Data: 
\code{\link{dv_data}},
\code{\link{em_data}},
\code{\link{g2g_data}},
\code{\link{mt_data}},
\code{\link{p2p_data_sim}()},
\code{\link{sq_data}}
}
\concept{Data}
\keyword{datasets}
