% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_report.R
\name{validation_report}
\alias{validation_report}
\title{Generate a Data Validation report in HTML}
\usage{
validation_report(
  data,
  meeting_data = NULL,
  hrvar = "Organization",
  path = "validation report",
  timestamp = TRUE
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{meeting_data}{An optional Meeting Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "Organization"}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file extension}.}

\item{timestamp}{Logical vector specifying whether to include a timestamp in the file name.
Defaults to TRUE.}
}
\description{
The function generates an interactive HTML report using
Standard Query data as an input. The report contains a checks on
Workplace Analytics query outputs, to provide diagnostic information
for the Analyst pre-analysis.

For your input data or meeting_data, please use the function \code{wpa::import_wpa()}
to import your csv query files into R. This function will standardize format
and prepare the data as input for this report.
}
\examples{
\dontrun{
validation_report(dv_data,
                  meeting_data = mt_data,
                  hrvar = "Organization")

}

}
\seealso{
Other Reports: 
\code{\link{IV_report}()},
\code{\link{capacity_report}()},
\code{\link{coaching_report}()},
\code{\link{collaboration_report}()},
\code{\link{connectivity_report}()},
\code{\link{generate_report}()}
}
\concept{Reports}
