% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellbeing_report.R
\name{wellbeing_report}
\alias{wellbeing_report}
\title{Wellbeing Report}
\usage{
wellbeing_report(
  wbq,
  hcq,
  hrvar = "Organization",
  mingroup = 5,
  output_file = "wellbeing_report.html"
)
}
\description{
Generate a wellbeing report
}
