% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workloads_dist.R
\name{workloads_dist}
\alias{workloads_dist}
\alias{workloads_distribution}
\title{Work Week Span distribution}
\usage{
workloads_dist(data, hrvar = "Organization", mingroup = 5, return = "plot")

workloads_distribution(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze Work Week Span distribution.
Returns a a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
workloads_dist(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Workloads: 
\code{\link{workloads_line}()},
\code{\link{workloads_summary}()},
\code{\link{workloads_trend}()}
}
\concept{Workloads}
