% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workloads_rank.R
\name{workloads_rank}
\alias{workloads_rank}
\title{Rank all groups across HR attributes for Work Week Span}
\usage{
workloads_rank(data, hrvar = extract_hr(data), mingroup = 5, return = "table")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{A character vector specifying what to return.
Valid values include "table" (default). Features are being considered for alternative return options but are currently
unavailable.}
}
\value{
When 'table' is passed in \code{return}, a summary table is returned as a data frame.
}
\description{
This function scans a standard query output for groups with high levels of Work Week Span.
Returns a table with a all of groups (across multiple HR attributes) ranked by work week span.
}
\details{
Uses the metric \code{Workweek_span}.
See \code{create_rank()} for applying the same analysis to a different metric.
}
\seealso{
Other Workloads: 
\code{\link{workloads_dist}()},
\code{\link{workloads_fizz}()},
\code{\link{workloads_line}()},
\code{\link{workloads_summary}()},
\code{\link{workloads_trend}()}
}
\concept{Workloads}
