% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workloads_summary.R
\name{workloads_summary}
\alias{workloads_summary}
\title{Work Week Span Summary}
\usage{
workloads_summary(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides an overview analysis of 'Work Week Span'.
Returns a bar plot showing average weekly utilization hours by default.
Additional options available to return a summary table.
}
\examples{
# Return a ggplot bar chart
workloads_summary(sq_data, hrvar = "LevelDesignation")

# Return a summary table
workloads_summary(sq_data, hrvar = "LevelDesignation", return = "table")
}
\seealso{
Other Workloads: 
\code{\link{workloads_dist}()},
\code{\link{workloads_fizz}()},
\code{\link{workloads_line}()},
\code{\link{workloads_trend}()}
}
\concept{Workloads}
