% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workloads_trend.R
\name{workloads_trend}
\alias{workloads_trend}
\title{Work Week Span Time Trend}
\usage{
workloads_trend(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of Work Week Span.
By default returns a week by week heatmap, highlighting the points in time with most activity.
Additional options available to return a summary table.
}
\examples{
## Run plot
workloads_trend(sq_data)

## Run table
workloads_trend(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Workloads: 
\code{\link{workloads_dist}()},
\code{\link{workloads_fizz}()},
\code{\link{workloads_line}()},
\code{\link{workloads_summary}()}
}
\concept{Workloads}
