% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_classify_pav.R
\name{workpatterns_classify_pav}
\alias{workpatterns_classify_pav}
\title{Classify working pattern personas using a rule based algorithm, using
the person-average volume-based (pav) method.}
\usage{
workpatterns_classify_pav(
  data,
  values = "percent",
  signals = "email",
  start_hour = "0900",
  end_hour = "1700",
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing data from the Hourly Collaboration query.}

\item{values}{Character vector to specify whether to return percentages
or absolute values in "data" and "plot". Valid values are "percent" (default)
and "abs".}

\item{signals}{Character vector to specify which collaboration metrics to use:
You may use "email" (default) for emails only, "IM" for Teams messages only,
or a combination of the two \code{c("email", "IM")}.}

\item{start_hour}{A character vector specifying starting hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying starting hours,
e.g. "1700"}

\item{return}{Character vector to specify what to return.
Valid options include "plot" (default), "data", "table" and "plot-area".
"plot" returns a bar plot, whilst "plot-area" returns an overlapping area plot.}
}
\description{
Apply a rule based algorithm to emails or instant messages sent by hour of day.
This uses a person-average volume-based (pav) method.
}
\examples{
\dontrun{
# Returns a plot by default
workpatterns_classify(em_data)

# Returning a table
workpatterns_classify(em_data, return = "table")
}

}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{personas_hclust}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
