% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_rank.R
\name{workpatterns_rank}
\alias{workpatterns_rank}
\title{Create a rank table of working patterns}
\usage{
workpatterns_rank(
  data,
  signals = c("email", "IM"),
  start_hour = "0900",
  end_hour = "1700",
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing hourly collaboration data.}

\item{signals}{Character vector to specify which collaboration metrics to use:
You may use "email" (default) for emails only, "IM" for Teams messages only,
or a combination of the two \code{c("email", "IM")}.}

\item{start_hour}{A character vector specifying starting hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying starting hours,
e.g. "1700"}

\item{return}{String specifying what to return. Defaults to "plot",
with options to return a summary table ("table").}
}
\description{
Takes in an Hourly Collaboration query and returns a count
table of working patterns, ranked from the most common to the
least.
}
\examples{
\dontrun{
workpatterns_rank(em_data)

workpatterns_rank(em_data, signals = "email")
}
}
