% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_to_afterhours.R
\name{zoom_to_afterhours}
\alias{zoom_to_afterhours}
\title{Take a clean Zoom participants file and compute after-hours metrics}
\usage{
zoom_to_afterhours(data, wowa_file, utc_offset = NULL)
}
\arguments{
\item{data}{data frame containing a clean Zoom participants file.}

\item{wowa_file}{data frame containing a Ways of Working Assessment flexible
query.}

\item{utc_offset}{string with the format 'hh:mm' to take into account the
time difference between the Zoom Admin's timezone and UTC.}
}
\description{
Internal function used within \code{zoom_to_pq()}, where a formatted Zoom
participants file is used to compute after-hours metrics.
}
