% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_to_pq.R
\name{zoom_to_pq}
\alias{zoom_to_pq}
\title{Convert a hashed Zoom participants data frame to a Person Query structure}
\usage{
zoom_to_pq(data, mq_key = "", wowa_file = NULL, utc_offset, return = "minimal")
}
\arguments{
\item{data}{data frame containing a hashed Zoom meeting participants dataset.}

\item{mq_key}{character vector containing the unique meeting subject lines
from the Standard Meeting Query. This is used for identifying whether calls
in Zoom are scheduled or unscheduled. If left blank, no matching is performed
and calls in the Zoom file are assumed to be unscheduled if the Meeting ID
and Creation Time Rules are satisfied.}

\item{wowa_file}{data frame containing a Ways of Working Assessment query.
Note that the \code{TimeZone} variable must be present in order for after-hours
metrics to be calculated correctly.}

\item{utc_offset}{string, with the format 'hh:mm' to represent the difference
of the Zoom Admin's timezone setting with respect to UTC. For example, you
should pass "08:00" to represent the Zoom Admin's timezone being 8 hours
ahead of UTC. For offsets which are \emph{behind} UTC, you may represent by
prefixing it with a minus sign, e.g. "-08:00", or directly deduct the value
from 24, e.g. "16:00".}

\item{return}{string determining what output to return. Must be one of the
following values:
\itemize{
\item \code{"minimal"}
\item \code{"full"}
\item \code{"afterhours"}
\item \code{"participants"}
}}
}
\description{
Convert a hashed Zoom participants level data frame to one with a Person
Query structure, where each row represents a person-week. A Standard Meeting
Query as well as a Person Query is required as inputs in order to perform
scheduled meeting and licensed population matching.
}
