% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_dist.R
\name{afterhours_dist}
\alias{afterhours_dist}
\alias{afterhours_distribution}
\title{After-Hours distribution}
\usage{
afterhours_dist(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(1, 2, 3)
)

afterhours_distribution(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(1, 2, 3)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{cut}{A vector specifying the cuts to use for the data,
accepting "default" or "range-cut" as character vector,
or a numeric value of length three to specify the exact breaks to use. e.g. c(1, 3, 5)}
}
\description{
Analyse the distribution of weekly after-hours collaboration time.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
}
\examples{
\dontrun{
## Return a plot
afterhours_dist(sq_data, hrvar = "Organization")

## Return a table
afterhours_dist(sq_data, hrvar = "Organization", return = "table")

## Return result with a custom specified breaks
afterhours_dist(sq_data, hrvar = "LevelDesignation", cut = c(4, 7, 9))
}
}
\seealso{
Other After-Hours: 
\code{\link{afterhours_rank}()},
\code{\link{afterhours_summary}()},
\code{\link{afterhours_trend}()}
}
\concept{After-Hours}
