% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_line.R
\name{afterhours_line}
\alias{afterhours_line}
\title{After-hours Collaboration Time Trend - Line Chart}
\usage{
afterhours_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of after-hours collaboration time, visualized as line charts.
By default returns a line chart for after-hours collaboration hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
See \code{create_line()} for applying the same analysis to a different metric.
}
\examples{

# Return a line plot
afterhours_line(sq_data, hrvar = "LevelDesignation")


# Return summary table
afterhours_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other After-Hours: 
\code{\link{afterhours_dist}()},
\code{\link{afterhours_fizz}()},
\code{\link{afterhours_rank}()},
\code{\link{afterhours_summary}()},
\code{\link{afterhours_trend}()}
}
\concept{After-Hours}
