% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_summary.R
\name{afterhours_summary}
\alias{afterhours_summary}
\alias{afterhours_sum}
\title{Summary of After-Hours Collaboration Hours}
\usage{
afterhours_summary(data, hrvar = "Organization", mingroup = 5, return = "plot")

afterhours_sum(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides an overview analysis of after-hours collaboration time.
Returns a bar plot showing average weekly after-hours collaboration hours by default.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
}
\examples{
# Return a ggplot bar chart
afterhours_summary(sq_data, hrvar = "LevelDesignation")

# Return a summary table
afterhours_summary(sq_data, hrvar = "LevelDesignation", return = "table")
}
\seealso{
Other After-Hours: 
\code{\link{afterhours_dist}()},
\code{\link{afterhours_fizz}()},
\code{\link{afterhours_line}()},
\code{\link{afterhours_rank}()},
\code{\link{afterhours_trend}()}
}
\concept{After-Hours}
