% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterhours_trend.R
\name{afterhours_trend}
\alias{afterhours_trend}
\title{After-Hours Time Trend}
\usage{
afterhours_trend(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of after-hours collaboration time.
By default returns a week by week heatmap, highlighting the points in time with most activity.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{After_hours_collaboration_hours}.
}
\examples{
## Run plot
afterhours_trend(sq_data)

## Run table
afterhours_trend(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other After-Hours: 
\code{\link{afterhours_dist}()},
\code{\link{afterhours_fizz}()},
\code{\link{afterhours_line}()},
\code{\link{afterhours_rank}()},
\code{\link{afterhours_summary}()}
}
\concept{After-Hours}
