% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{camel_clean}
\alias{camel_clean}
\title{Convert "CamelCase" to "Camel Case"}
\usage{
camel_clean(string)
}
\arguments{
\item{string}{A string vector in 'CamelCase' format to format}
}
\value{
Returns a formatted string.
}
\description{
Convert a text string from the format "CamelCase" to "Camel Case".
This is used for converting variable names such as
"LevelDesignation" to "Level Designation" for the purpose
of prettifying plot labels.
}
\examples{
camel_clean("NoteHowTheStringIsFormatted")

}
\seealso{
Other Support: 
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{is_date_format}()},
\code{\link{maxmin}()},
\code{\link{md2html}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{plot_WOE}()},
\code{\link{read_preamble}()},
\code{\link{rgb2hex}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{tstamp}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}
}
\concept{Support}
