% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_dist.R, R/create_dist.R
\name{collaboration_dist}
\alias{collaboration_dist}
\alias{collaboration_distribution}
\title{Collaboration Hours distribution}
\usage{
collaboration_dist(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(15, 20, 25)
)

collaboration_distribution(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(15, 20, 25)
)

collaboration_distribution(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(15, 20, 25)
)
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{cut}{A numeric vector of length three to specify the breaks for the distribution,
e.g. c(10, 15, 20)}
}
\description{
Analyze the distribution of Collaboration Hours.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\details{
Uses the metric \code{Collaboration_hours}.
}
\examples{
## Return a plot
collaboration_dist(sq_data, hrvar = "Organization")

## Return a table
collaboration_dist(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Collaboration: 
\code{\link{collaboration_area}()},
\code{\link{collaboration_fizz}()},
\code{\link{collaboration_line}()},
\code{\link{collaboration_rank}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()},
\code{\link{create_dist}()},
\code{\link{meeting_trend}()}
}
\concept{Collaboration}
