% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_rank.R
\name{collaboration_rank}
\alias{collaboration_rank}
\alias{collab_rank}
\title{Collaboration Ranking}
\usage{
collaboration_rank(
  data,
  hrvar = extract_hr(data),
  mingroup = 5,
  return = "table"
)

collab_rank(data, hrvar = extract_hr(data), mingroup = 5, return = "table")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{A character vector specifying what to return.
Valid values include "table" (default). Features are being considered for alternative return options but are currently
unavailable.}
}
\value{
When 'table' is passed in \code{return}, a summary table is returned as a data frame.
}
\description{
This function scans a standard query output for groups with high levels of 'Weekly Digital Collaboration'.
Returns a table with a all of groups (across multiple HR attributes) ranked by hours of digital collaboration.
}
\details{
Uses the metric \code{Collaboration_hours}.
See \code{create_rank()} for applying the same analysis to a different metric.
}
\seealso{
Other Collaboration: 
\code{\link{collaboration_area}()},
\code{\link{collaboration_dist}()},
\code{\link{collaboration_fizz}()},
\code{\link{collaboration_line}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()},
\code{\link{meeting_trend}()}
}
\concept{Collaboration}
