% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dist.R
\name{create_dist}
\alias{create_dist}
\title{Horizontal 100 percent stacked bar plot for any metric}
\usage{
create_dist(
  data,
  metric,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(15, 20, 25),
  dist_colours = c("#facebc", "#fcf0eb", "#b4d5dd", "#bfe5ee"),
  unit = "hours"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{String containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}

\item{cut}{A numeric vector of length three to specify the breaks for the distribution,
e.g. c(10, 15, 20)}

\item{dist_colours}{A character vector of length four to specify colour
codes for the stacked bars.}

\item{unit}{See \code{cut_hour()}.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A stacked bar plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Provides an analysis of the distribution of a selected metric.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
# Return a plot
create_dist(sq_data, metric = "Collaboration_hours", hrvar = "Organization")

# Return a table
create_dist(sq_data, metric = "Collaboration_hours", hrvar = "Organization", return = "table")
}
\seealso{
Other Flexible: 
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
