% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line.R
\name{create_line}
\alias{create_line}
\title{Time Trend - Line Chart for any metric}
\usage{
create_line(
  data,
  metric,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of a selected metric, visualised as line charts.
By default returns a line chart for the defined metric,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\details{
This is a general purpose function that powers all the functions
in the package that produce faceted line plots.
}
\examples{
sq_data \%>\% create_line(metric = "Email_hours", return = "plot")

sq_data \%>\% create_line(metric = "Collaboration_hours", return = "plot")

sq_data \%>\% create_line(metric = "Workweek_span", hrvar = "LevelDesignation")
}
\seealso{
Other Flexible: 
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
