% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{Create a sankey chart from a two-column count table}
\usage{
create_sankey(data, var1, var2, count = "n")
}
\arguments{
\item{data}{Data frame of the long count table.}

\item{var1}{String containing the name of the variable to be shown on the
left.}

\item{var2}{String containing the name of the variable to be shown on the
right.}

\item{count}{String containing the name of the count variable.}
}
\value{
A 'sankeyNetwork' and 'htmlwidget' object containing a two-tier
sankey plot. The output can be saved locally with
\code{htmlwidgets::saveWidget()}.
}
\description{
Create a 'networkD3' style sankey chart based on a long count table
with two variables. The input data should have three columns, where
each row is a unique group:
\enumerate{
\item Variable 1
\item Variable 2
\item Count
}
}
\examples{
\donttest{
sq_data \%>\%
  dplyr::count(Organization, FunctionType) \%>\%
  create_sankey(var1 = "Organization", var2 = "FunctionType")
}

}
\seealso{
Other Visualization: 
\code{\link{afterhours_dist}()},
\code{\link{afterhours_fizz}()},
\code{\link{afterhours_line}()},
\code{\link{afterhours_rank}()},
\code{\link{afterhours_summary}()},
\code{\link{afterhours_trend}()},
\code{\link{collaboration_area}()},
\code{\link{collaboration_dist}()},
\code{\link{collaboration_fizz}()},
\code{\link{collaboration_line}()},
\code{\link{collaboration_rank}()},
\code{\link{collaboration_sum}()},
\code{\link{collaboration_trend}()},
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()},
\code{\link{email_dist}()},
\code{\link{email_fizz}()},
\code{\link{email_line}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()},
\code{\link{email_trend}()},
\code{\link{external_network_plot}()},
\code{\link{hr_trend}()},
\code{\link{hrvar_count}()},
\code{\link{internal_network_plot}()}

Other Flexible: 
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_hist}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
\concept{Visualization}
