% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_hour.R
\name{cut_hour}
\alias{cut_hour}
\title{Convert a numeric variable for hours into categorical}
\usage{
cut_hour(metric, cuts, unit = "hours")
}
\arguments{
\item{metric}{A numeric variable representing hours.}

\item{cuts}{A numeric variable of minimum length 3 to represent the
cut points required.}

\item{unit}{String to specify the unit of the labels. Defaults to "hours".}
}
\description{
Supply a numeric variable, e.g. \code{Collaboration_hours},
and the function returns a character vector
}
\details{
This is used within \code{create_dist()} for numeric to categorical conversion.
}
\examples{
# Direct use
cut_hour(1:30, cuts = c(15, 20, 25))

# Use on a query
cut_hour(sq_data$Collaboration_hours, cuts = c(10, 15, 20))

}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{extract_date_range}()},
\code{\link{rgb2hex}()}
}
\concept{Support}
