% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_dist.R
\name{email_dist}
\alias{email_dist}
\title{Distribution of Email Hours as a 100\% stacked bar}
\usage{
email_dist(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(5, 10, 15)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character string specifying what to return. Valid inputs are:
\itemize{
\item "plot": ggplot object. A stacked bar plot for the metric.
\item "table": data frame. A summary table for the metric.
}}

\item{cut}{A numeric vector of length three to specify the breaks for the distribution,
e.g. c(10, 15, 20)}
}
\description{
Analyze Email Hours distribution.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
## Return a plot
email_dist(sq_data, hrvar = "Organization")

## Return a table
email_dist(sq_data, hrvar = "Organization", return = "table")

## Return result with a custom specified breaks
email_dist(sq_data, hrvar = "LevelDesignation", cut = c(4, 7, 9))

}
\seealso{
Other Emails: 
\code{\link{email_fizz}()},
\code{\link{email_line}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()},
\code{\link{email_trend}()}
}
\concept{Emails}
