% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_fizz.R
\name{email_fizz}
\alias{email_fizz}
\title{Distribution of Email Hours (Fizzy Drink plot)}
\usage{
email_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze weekly email hours distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
email_fizz(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Emails: 
\code{\link{email_dist}()},
\code{\link{email_line}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()},
\code{\link{email_trend}()}
}
\concept{Emails}
