% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_line.R
\name{email_line}
\alias{email_line}
\title{Email Time Trend - Line Chart}
\usage{
email_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of email time, visualised as line charts.
By default returns a line chart for email hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\examples{

# Return a line plot
email_line(sq_data, hrvar = "LevelDesignation")


# Return summary table
email_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Emails: 
\code{\link{email_dist}()},
\code{\link{email_fizz}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()},
\code{\link{email_trend}()}
}
\concept{Emails}
