% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_trend.R
\name{email_trend}
\alias{email_trend}
\title{Email Hours Time Trend}
\usage{
email_trend(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of email time.
By default returns a week by week heatmap, highlighting the points in time with most activity.
Additional options available to return a summary table.
}
\details{
Uses the metric \code{Email_hours}.
}
\seealso{
Other Emails: 
\code{\link{email_dist}()},
\code{\link{email_fizz}()},
\code{\link{email_line}()},
\code{\link{email_rank}()},
\code{\link{email_summary}()}
}
\concept{Emails}
