% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_network_plot.R
\name{external_network_plot}
\alias{external_network_plot}
\title{Plot the external network metrics for a HR variable}
\usage{
external_network_plot(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  bubble_size = c(1, 5)
)
}
\arguments{
\item{data}{Person Query as a dataframe including date column named "Date"
This function assumes the data format is MM/DD/YYYY as is standard in a WpA query output.}

\item{hrvar}{WpA variable for an HR variable to group networks by
For example, "Layer"}

\item{mingroup}{Numeric vector for minimum group size for aggregation}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{bubble_size}{A numeric vector of length two to specify the size range of the bubbles}
}
\description{
Plot the external network metrics for a HR variable.
}
\examples{
\dontrun{external_network_plot(sq_data)
}

}
\seealso{
Other Connectivity: 
\code{\link{internal_network_plot}()}
}
\concept{Connectivity}
