% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_network_plot.R
\name{external_network_plot}
\alias{external_network_plot}
\title{Plot External Network Breadth and Size as a scatter plot}
\usage{
external_network_plot(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  bubble_size = c(1, 8)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{bubble_size}{A numeric vector of length two to specify the size range of the bubbles}
}
\description{
Plot the external network metrics for a HR variable as a scatter plot, showing
External Network Breadth as the vertical axis and External Network Size as the
horizontal axis.
}
\examples{
# Return plot
sq_data \%>\% external_network_plot(return = "plot")

}
\seealso{
Other Connectivity: 
\code{\link{internal_network_plot}()}
}
\concept{Connectivity}
