% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_ch_ratio.R
\name{flag_ch_ratio}
\alias{flag_ch_ratio}
\title{Flag unusual high collaboration hours to after-hours collaboration
hours ratio}
\usage{
flag_ch_ratio(data, threshold = c(1, 30), return = "message")
}
\arguments{
\item{data}{A data frame containing a Person Query.}

\item{threshold}{Numeric value specifying the threshold for flagging.
Defaults to 30.}

\item{return}{String to specify what to return. Options include:
\itemize{
\item \code{"message"}
\item \code{"text"}
\item \code{"data"}
}}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"message"}: message in the console containing diagnotic summary
\item \code{"text"}: string containing diagnotic summary
\item \code{"data"}: data frame. Person-level data with flags on unusually high or
low ratios
}
}
\description{
This function flags persons who have an unusual ratio
of collaboration hours to after-hours collaboration hours.
Returns a character string by default.
}
\examples{
flag_ch_ratio(sq_data)


data.frame(PersonId = c("Alice", "Bob"),
           Collaboration_hours = c(30, 0.5),
           After_hours_collaboration_hours = c(0.5, 30)) \%>\%
  flag_ch_ratio()

}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{extract_hr}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hr_trend}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
