% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_em_ratio.R
\name{flag_em_ratio}
\alias{flag_em_ratio}
\title{Flag Persons with unusually high Email Hours to Emails Sent ratio}
\usage{
flag_em_ratio(data, threshold = 1, return = "text")
}
\arguments{
\item{data}{A data frame containing a Person Query.}

\item{threshold}{Numeric value specifying the threshold for flagging. Defaults to 1.}

\item{return}{Character vector specifying what to return. Defaults to "text", with
valid options to return a data frame ("data").}
}
\description{
This function flags persons who have an unusual ratio
of email hours to emails sent. If the ratio between Email Hours and
Emails Sent is greater than the threshold, then observations tied to
a \code{PersonId} is flagged as unusual.
}
\examples{
flag_em_ratio(sq_data)

}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hr_trend}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
