% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outlooktime.R
\name{flag_outlooktime}
\alias{flag_outlooktime}
\title{Flag unusual outlook time settings for work day start and end time}
\usage{
flag_outlooktime(data, threshold = c(4, 15), return = "message")
}
\arguments{
\item{data}{A data frame containing a Person Query.}

\item{threshold}{A numeric vector of length two, specifying the hour threshold for flagging.
Defaults to c(4, 15).}

\item{return}{String to specify what to return.
Valid options include "text" (default), "message", and "data".}
}
\description{
This function flags unusual outlook calendar settings for
start and end time of work day.
}
\examples{
flag_outlooktime(dv_data)
}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{hr_trend}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
