% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2g_data.R
\docType{data}
\name{g2g_data}
\alias{g2g_data}
\title{Sample Group-to-Group dataset}
\format{
A data frame with 3517 rows and 7 variables:
\describe{
\item{TimeInvestors_Organization}{ }
\item{Collaborators_Organization}{ }
\item{Date}{ }
\item{Meetings}{ }
\item{Meeting_hours}{ }
\item{Email_hours}{ }
\item{Collaboration_hours}{ }

...
}
}
\source{
\url{https://workplaceanalytics-demo.office.com/en-us/Home}
}
\usage{
g2g_data
}
\description{
A demo dataset generated from a Group-to-Group Query from Workplace Analytics.
}
\seealso{
Other Data: 
\code{\link{dv_data}},
\code{\link{em_data}},
\code{\link{p2p_data_sim}()}

Other Network: 
\code{\link{external_network_plot}()},
\code{\link{internal_network_plot}()},
\code{\link{network_g2g}()},
\code{\link{network_leiden}()},
\code{\link{network_louvain}()},
\code{\link{network_p2p}()},
\code{\link{p2p_data_sim}()}
}
\concept{Data}
\concept{Network}
\keyword{datasets}
