% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvar_count.R
\name{hrvar_count}
\alias{hrvar_count}
\alias{analysis_scope}
\title{Create a count of distinct people in a specified HR variable}
\usage{
hrvar_count(data, hrvar = "Organization", return = "plot")

analysis_scope(data, hrvar = "Organization", return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation".
If a vector with more than one value is provided, the HR attributes are automatically
concatenated.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
This function enables you to create a count of the distinct people
by the specified HR attribute.The default behaviour is to return a
bar chart as typically seen in 'Analysis Scope'.
}
\examples{
# Return a bar plot
hrvar_count(sq_data, hrvar = "LevelDesignation")

# Return a summary table
hrvar_count(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other General: 
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_scatter}()},
\code{\link{cut_hour}()},
\code{\link{extract_hr}()},
\code{\link{identify_outlier}()}

Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hrvar_count_all}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
\concept{General}
