% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvar_count_all.R
\name{hrvar_count_all}
\alias{hrvar_count_all}
\title{Create count of distinct fields and percentage of employees with NAs for all HR variables}
\usage{
hrvar_count_all(
  data,
  n_var = 50,
  return = "message",
  threshold = 100,
  maxna = 20
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{n_var}{number of HR variables to include in report as rows. Default is set to 10 HR variables.}

\item{return}{String to specify what to return}

\item{threshold}{The max number of unique values allowed for any attribute. Default is 100.}

\item{maxna}{The max percentage of NAs allowable for any column. Default is 20.}
}
\value{
Returns an error message by default, where 'text' is passed in \code{return}.
When 'table' is passed, a summary table listing the number of distinct fields and percentage of NAs for the specified number of HR attributes will be returned.
when 'message' is passed, outputs a message indicating which values are beyond the specified thresholds.
}
\description{
This function enables you to create a summary table to validate organizational data. This table will provide a summary of the data found
in the WpA Sources page.
This function will return a summary table with the count of distinct fields per HR attribute and the percentage of
employees with NAs for that attribute.
See hrvar_count function for more detail on the specific HR attribute of interest.
}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hrvar_count}()},
\code{\link{identify_holidayweeks}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
