% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_holidayweeks.R
\name{identify_holidayweeks}
\alias{identify_holidayweeks}
\title{Identify Holiday Weeks}
\usage{
identify_holidayweeks(data, sd = 1, return = "message")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{sd}{The standard deviation below the mean for collaboration hours that should define an outlier week. Enter a positive number. Default is 1 standard deviation.}

\item{return}{String to specify what to return. Returns a message by default (\code{return} = "message").
Valid options are:
\itemize{
\item When 'message' is passed, a message is printed identifying holiday weeks.
\item When 'data' is passed, a dataset with outlier weeks flagged in a new column is returned as a dataframe.
\item When 'data_cleaned' is passed, a dataset with outlier weeks removed is returned as a dataframe.
\item when 'data_dirty' is passed, a dataset with only outlier weeks is returned as a dataframe.
\item when 'plot' is passed, a pot with holiday weeks highlighted is returned as a dataframe.
}}
}
\value{

}
\description{
This function scans a standard query output for weeks where collaboration hours is far outside the mean.
Returns a list of weeks that appear to be holiday weeks and optionally an edited dataframe with outliers removed.
By default, missing values are excluded.

As best practice, run this function prior to any analysis to remove atypical collaboration weeks from your dataset.
}
\seealso{
Other Data Validation: 
\code{\link{check_query}()},
\code{\link{flag_ch_ratio}()},
\code{\link{flag_em_ratio}()},
\code{\link{flag_extreme}()},
\code{\link{flag_outlooktime}()},
\code{\link{hrvar_count_all}()},
\code{\link{hrvar_count}()},
\code{\link{identify_inactiveweeks}()},
\code{\link{identify_nkw}()},
\code{\link{identify_outlier}()},
\code{\link{identify_privacythreshold}()},
\code{\link{identify_query}()},
\code{\link{identify_tenure}()},
\code{\link{remove_outliers}()},
\code{\link{subject_validate_report}()},
\code{\link{subject_validate}()},
\code{\link{track_HR_change}()}
}
\concept{Data Validation}
