% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_shifts_wp.R
\name{identify_shifts_wp}
\alias{identify_shifts_wp}
\title{Identify shifts based on binary activity}
\usage{
identify_shifts_wp(
  data,
  signals = c("email", "IM"),
  active_threshold = 1,
  start_hour = 9,
  end_hour = 17,
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing data from the Hourly Collaboration query.}

\item{signals}{Character vector to specify which collaboration metrics to
use:
\itemize{
\item a combination of signals, such as \code{c("email", "IM")} (default)
\item \code{"email"} for emails only
\item \code{"IM"} for Teams messages only
\item \code{"unscheduled_calls"} for Unscheduled Calls only
\item \code{"meetings"} for Meetings only
}}

\item{active_threshold}{A numeric value specifying the minimum number of
signals to be greater than in order to qualify as \emph{active}. Defaults to 0.}

\item{start_hour}{A character vector specifying start hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying finish hours,
e.g. "1700"}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
\item \code{"data"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"plot"}: ggplot object. A bar plot for the weekly count of shifts.
\item \code{"table"}: data frame. A summary table for the count of shifts.
\item \verb{"data}: data frame. Input data appended with the following columns:
\itemize{
\item \code{Start}
\item \code{End}
\item \code{DaySpan}
\item \code{Shifts}
}
}
}
\description{
This function uses the Hourly Collaboration query and computes binary
activity to identify the 'behavioural' work shift. This is a distinct method
to \code{identify_shifts()}, which instead uses outlook calendar settings for
start and end time of work day to identify work shifts. The two methods can
be compared to gauge the accuracy of existing Outlook settings.
}
\examples{
# Return plot
em_data \%>\% identify_shifts_wp()

# Return table
em_data \%>\% identify_shifts_wp(return = "table")

}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify_pav}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
