% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_network_plot.R
\name{internal_network_plot}
\alias{internal_network_plot}
\title{Plot Internal Network Breadth and Size as a scatter plot}
\usage{
internal_network_plot(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  bubble_size = c(1, 8)
)
}
\arguments{
\item{data}{Person Query as a dataframe including date column named "Date"
This function assumes the data format is MM/DD/YYYY as is standard in a WpA query output.}

\item{hrvar}{WpA variable for an HR variable to group networks by
For example, "Layer"}

\item{mingroup}{Numeric vector for minimum group size for aggregation}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}

\item{bubble_size}{A numeric vector of length two to specify the size range of the bubbles}
}
\description{
Plot the internal network metrics for a HR variable as a scatter plot, showing
Internal Network Breadth as the vertical axis and Internal Network Size as the
horizontal axis.
}
\examples{
\dontrun{
internal_network_plot(sq_data)
}

}
\seealso{
Other Connectivity: 
\code{\link{external_network_plot}()}
}
\concept{Connectivity}
