% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_format.R
\name{is_date_format}
\alias{is_date_format}
\title{Identify whether string is a date format}
\usage{
is_date_format(string)
}
\arguments{
\item{string}{Character string to test whether is a date format.}
}
\value{
logical value indicating whether the string is a date format.
}
\description{
This function uses regular expression to determine whether a string is of the
format \code{"mdy"}, separated by \code{"-"}, \code{"/"}, or \code{"."}, returning a logical
vector.
}
\examples{
is_date_format("1/5/2020")

}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{maxmin}()},
\code{\link{md2html}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{plot_WOE}()},
\code{\link{read_preamble}()},
\code{\link{rgb2hex}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{tstamp}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}
}
\concept{Support}
