% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_format.R
\name{is_date_format}
\alias{is_date_format}
\title{Identify whether string is a date format}
\usage{
is_date_format(string)
}
\arguments{
\item{string}{Character string to test whether is a date format.}
}
\description{
This function uses regular expression to determine whether
a string is of the format "mdy", separated by "-", "/",
or ".", returning a logical vector.
}
\examples{
\dontrun{
is_date_format("1/5/2020")
}

}
