% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_dist.R
\name{meeting_dist}
\alias{meeting_dist}
\title{Distribution of Meeting Hours as a 100\% stacked bar}
\usage{
meeting_dist(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot",
  cut = c(5, 10, 15)
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character
vector, defaults to "Organization" but accepts any character vector, e.g.
"LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size, defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}

\item{cut}{A numeric vector of length three to specify the breaks for the
distribution,
e.g. c(10, 15, 20)}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"plot"}: ggplot object. A stacked bar plot for the metric.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
Analyze Meeting Hours distribution.
Returns a stacked bar plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
# Return plot
meeting_dist(sq_data, hrvar = "Organization")

# Return summary table
meeting_dist(sq_data, hrvar = "Organization", return = "table")

# Return result with a custom specified breaks
meeting_dist(sq_data, hrvar = "LevelDesignation", cut = c(4, 7, 9))

}
\seealso{
Other Meeting Culture: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()}
}
\concept{Meeting Culture}
