% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_line.R
\name{meeting_line}
\alias{meeting_line}
\title{Meeting Time Trend - Line Chart}
\usage{
meeting_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of meeting time, visualised as line charts.
By default returns a line chart for meeting hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\examples{

## Return a line plot
meeting_line(sq_data, hrvar = "LevelDesignation")


## Return a table
meeting_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Meetings: 
\code{\link{meeting_fizz}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_summary}()},
\code{\link{meeting_trend}()},
\code{\link{meetingtype_summary}()}
}
\concept{Meetings}
