% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_skim.R
\name{meeting_skim}
\alias{meeting_skim}
\title{Produce a skim summary of meeting hours}
\usage{
meeting_skim(data, return = "message")
}
\arguments{
\item{data}{A standard person query data in the form of a data frame.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"message"}
\item \code{"text"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return} argument:
\itemize{
\item \code{"message"}: message in console.
\item \code{"text"}: string.
\item \code{"table"}: data frame.
}
}
\description{
This function returns a skim summary in the console
when provided a standard query in the input.
}
\examples{
meeting_skim(sq_data)

}
\seealso{
Other Meetings: 
\code{\link{meeting_dist}()},
\code{\link{meeting_extract}()},
\code{\link{meeting_fizz}()},
\code{\link{meeting_line}()},
\code{\link{meeting_quality}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_summary}()},
\code{\link{meeting_tm_report}()},
\code{\link{meeting_trend}()},
\code{\link{meetingtype_dist}()},
\code{\link{meetingtype_dist_ca}()},
\code{\link{meetingtype_dist_mt}()},
\code{\link{meetingtype_summary}()}
}
\concept{Meetings}
