% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_summary.R
\name{meeting_summary}
\alias{meeting_summary}
\alias{meeting_sum}
\title{Meeting Summary}
\usage{
meeting_summary(data, hrvar = "Organization", mingroup = 5, return = "plot")

meeting_sum(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides an overview analysis of weekly meeting hours.
Returns a bar plot showing average weekly meeting hours by default.
Additional options available to return a summary table.
}
\examples{
# Return a ggplot bar chart
meeting_summary(sq_data, hrvar = "LevelDesignation")

# Return a summary table
meeting_summary(sq_data, hrvar = "LevelDesignation", return = "table")
}
\seealso{
Other Meetings: 
\code{\link{meeting_fizz}()},
\code{\link{meeting_line}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_trend}()},
\code{\link{meetingtype_summary}()}
}
\concept{Meetings}
