% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetingtype_dist.R
\name{meetingtype_dist}
\alias{meetingtype_dist}
\title{Distribution of Meeting Types by number of Attendees and Duration}
\usage{
meetingtype_dist(data, hrvar = NULL, mingroup = 5, return = "plot")
}
\arguments{
\item{data}{Data frame. If a meeting query, must contain the variables \code{Attendee} and \code{DurationHours}.}

\item{hrvar}{Character string to specify the HR attribute to split the data by.
Note that this is only applicable if a Collaboration Assessment query is passed to the function. If a Meeting Query
is passed instead, this argument is ignored.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.
Only applicable when using a Collaboration Assessment query.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Calculate the hour distribution of internal meeting types.
This is a wrapper around \code{meetingtype_dist_mt()} and
\code{meetingtype_dist_ca()}, depending on whether a Meeting Query or a Collaboration Assessment Query
is passed as an input.
}
