% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetingtype_dist_mt.R
\name{meetingtype_dist_mt}
\alias{meetingtype_dist_mt}
\title{Meeting Type Distribution (Meeting Query)}
\usage{
meetingtype_dist_mt(data, return = "plot")
}
\arguments{
\item{data}{Meeting Query data frame. Must contain the variables \code{Attendee} and \code{DurationHours}}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Calculate the hour distribution of internal meeting types,
using a Meeting Query with core WpA variables as an input.
}
