% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_describe.R
\name{network_describe}
\alias{network_describe}
\title{Uncover HR attributes which best represent a population for a Person to Person query}
\usage{
network_describe(
  data,
  hrvar = c("Organization", "LevelDesignation", "FunctionType")
)
}
\arguments{
\item{data}{Data frame for a person to person query.}

\item{hrvar}{Character vector of length 3 containing the HR attributes to be used.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("experimental")}
Returns a data frame that gives a percentage of the group combinations that best represent
the population provided. Uses a person to person query.
}
\author{
Tannaz Sattari Tabrizi \href{mailto:Tannaz.Sattari@microsoft.com}{Tannaz.Sattari@microsoft.com}
}
