% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_leiden.R
\name{network_leiden}
\alias{network_leiden}
\title{Implement the Leiden community detection on a Person to Person network query}
\usage{
network_leiden(
  data,
  hrvar,
  bg_fill = "#000000",
  font_col = "#FFFFFF",
  algorithm = "mds",
  node_alpha = 0.8,
  res = 0.5,
  desc_hrvar = c("Organization", "LevelDesignation", "FunctionType"),
  return
)
}
\arguments{
\item{data}{Data frame containing a Person to Person query.}

\item{hrvar}{String containing the HR attribute to be matched in the dataset.}

\item{bg_fill}{String to specify background fill colour.}

\item{font_col}{String to specify font and link colour.}

\item{algorithm}{String to specify the node placement algorithm to be used. Defaults to "fr" for the force-directed
algorithm of Fruchterman and Reingold. See \url{https://rdrr.io/cran/ggraph/man/layout_tbl_graph_igraph.html} for a
full list of options.}

\item{node_alpha}{A numeric value between 0 and 1 to specify the transparency of the nodes.}

\item{res}{Resolution parameter to be passed to \code{leiden::leiden()}. Defaults to 0.5.}

\item{return}{String specifying what output to return. Valid return options include:
\itemize{
\item 'plot-leiden': return a network plot coloured by leiden communities.
\item 'plot-hrvar': return a network plot coloured by HR attribute.
\item 'plot-sankey': return a sankey plot combining communities and HR attribute.
\item 'table': return a vertex summary table with counts in communities and HR attribute.
\item 'data': return a vertex data file that matches vertices with communities and HR attributes.
\item 'describe': returns a list of data frames which describe each of the identified communities.
}}
}
\description{
Take a P2P network query and implement the Leiden community detection method. To run
this function, you will require all the pre-requisites of the \strong{leiden} package installed,
which includes Python and \strong{reticulate}.
}
