% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_louvain.R
\name{network_louvain}
\alias{network_louvain}
\title{Implement the Louvain community detection on a Person to Person network query}
\usage{
network_louvain(
  data,
  hrvar = "Organization",
  bg_fill = "#000000",
  font_col = "#FFFFFF",
  node_alpha = 0.8,
  algorithm = "mds",
  path = "network_p2p_louvain",
  desc_hrvar = c("Organization", "LevelDesignation", "FunctionType"),
  return = "plot-louvain",
  size_threshold = 5000
)
}
\arguments{
\item{data}{Data frame containing a Person to Person Network query. Note that this function is
computationally intensive and may take a noticeably longer time to process beyond 5000 rows.}

\item{hrvar}{String containing the HR attribute to be matched in the dataset.
Defaults to "Organization".}

\item{bg_fill}{String to specify background fill colour.}

\item{font_col}{String to specify font and link colour.}

\item{node_alpha}{A numeric value between 0 and 1 to specify the transparency of the nodes.}

\item{algorithm}{String to specify the node placement algorithm to be used. Defaults to "mds" to perform
a multidimensional scaling of nodes using a shortest path, which is also a deterministic method.
See \url{https://rdrr.io/cran/ggraph/man/layout_tbl_graph_igraph.html} for a full list of options.}

\item{path}{File path for saving the PDF output. Defaults to "network_p2p_louvain".
Since the network outputs are computationally intensive, the default behaviour is to save time by
saving the plot output directly as a PDF in the specified path. To override this behaviour and return
a plot object instead, you can pass \code{NULL} to \code{path}. What is passed to \code{path} makes no difference
if returning anything other than "plot-louvain" or "plot-hrvar".}

\item{desc_hrvar}{Character vector of length 3 containing the HR attributes to use when returning the
"describe" output. See \code{network_describe()}.}

\item{return}{String specifying what output to return. Defaults to "plot-louvain".
Valid return options include:
\itemize{
\item 'plot-louvain': return a network plot coloured by louvain communities, saving a PDF to \code{path}.
\item 'plot-hrvar': return a network plot coloured by HR attribute, saving a PDF to \code{path}.
\item 'plot-sankey': return a sankey plot combining communities and HR attribute.
\item 'table': return a vertex summary table with counts in communities and HR attribute.
\item 'data': return a vertex data file that matches vertices with communities and HR attributes.
\item 'describe': returns a list of data frames which describe each of the identified communities.
\item 'network': return igraph object.
}}

\item{size_threshold}{Numeric value representing the maximum number of edges before \code{network_leiden()}
switches to use a more efficient, but less elegant plotting method (native igraph). Defaults to 5000.
Set as \code{0} to co-erce to a fast plotting method every time, and \code{Inf} to always use the default plotting
method.}
}
\description{
Take a P2P network query and implement the Louvain community detection method. The
\strong{igraph} implementation of the Louvain method is used.
}
