% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_p2p.R
\name{network_p2p}
\alias{network_p2p}
\title{Create a network plot with the person-to-person query}
\usage{
network_p2p(
  data,
  hrvar,
  return = "plot",
  bg_fill = "#000000",
  font_col = "#FFFFFF",
  ...
)
}
\arguments{
\item{data}{Data frame containing a person-to-person query.}

\item{hrvar}{String containing the label for the HR attribute.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot", "network" and "table". "network" returns the \code{network}
object used to generate the network plot.}

\item{...}{Additional arguments to pass to \code{GGally::ggnet2()}}
}
\description{
Pass a data frame containing a person-to-person query and return a network
plot.
Automatically handles "Collaborators_within_group" and "Other_collaborators" within query data.
}
