% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_line.R
\name{one2one_line}
\alias{one2one_line}
\title{Manager 1:1 Time Trend - Line Chart}
\usage{
one2one_line(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a ggplot object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides a week by week view of 1:1 time with managers, visualised as line charts.
By default returns a line chart for 1:1 meeting hours,
with a separate panel per value in the HR attribute.
Additional options available to return a summary table.
}
\examples{

## Return a line plot
one2one_line(sq_data, hrvar = "LevelDesignation")


## Return a table
one2one_line(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_rank}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}
}
\concept{Managerial Relations}
