% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2one_rank.R
\name{one2one_rank}
\alias{one2one_rank}
\title{Manager 1:1 Time Ranking}
\usage{
one2one_rank(data, hrvar = extract_hr(data), mingroup = 5, return = "table")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{A list of HR Variables to consider in the scan.
Defaults to all HR attributes identified.}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{A character vector specifying what to return.
Valid values include "table" (default). Features are being considered for alternative return options but are currently
unavailable.}
}
\value{
When 'table' is passed in \code{return}, a summary table is returned as a data frame.
}
\description{
This function scans a standard query output for groups with high levels of 'Manager 1:1 Time'.
Returns a table with a all of groups (across multiple HR attributes) ranked by hours of digital collaboration.
}
\details{
Uses the metric \code{Meeting_hours_with_manager_1_on_1}.
See \code{create_rank()} for applying the same analysis to a different metric.
}
\seealso{
Other Managerial Relations: 
\code{\link{mgrcoatt_dist}()},
\code{\link{mgrrel_matrix}()},
\code{\link{one2one_dist}()},
\code{\link{one2one_fizz}()},
\code{\link{one2one_line}()},
\code{\link{one2one_sum}()},
\code{\link{one2one_trend}()}
}
\concept{Managerial Relations}
