% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personas_hclust.R
\name{personas_hclust}
\alias{personas_hclust}
\title{Create hierarchical clusters of selected metrics using a Person query}
\usage{
personas_hclust(data, metrics, k = 4, return = "plot")
}
\arguments{
\item{data}{A data frame containing \code{PersonId} and selected metrics for clustering.}

\item{metrics}{Character vector containing names of metrics to use for clustering. See examples section.}

\item{k}{Numeric vector to specify the \code{k} number of clusters to cut by.}

\item{return}{Character vector to specify what to return.
Valid options include:
\itemize{
\item "plot": return
\item "data": return raw data with clusters appended
\item "table": return summary table for identified clusters
\item "hclust": return hierarchical model (an \code{hclust} object) generated by the function.
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Apply hierarchical clustering to selected metrics. Person averages are computed prior to clustering.
The hierarchical clustering uses cosine distance and the ward.D method
of agglomeration.
}
\examples{
# Return plot
personas_hclust(sq_data,
                metrics = c("Collaboration_hours", "Workweek_span"),
                k = 4)

# Return summary table
personas_hclust(sq_data,
                metrics = c("Collaboration_hours", "Workweek_span"),
                k = 4,
                return = "table")

\donttest{
# Return data with clusters appended
personas_hclust(sq_data,
                metrics = c("Collaboration_hours", "Workweek_span"),
                k = 4,
                return = "data")
}

}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{identify_shifts_wp}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify_pav}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
