% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IV.R
\name{plot_WOE}
\alias{plot_WOE}
\title{Plot WOE graphs with an IV object}
\usage{
plot_WOE(IV, predictor)
}
\arguments{
\item{IV}{IV object created with 'Information'.}

\item{predictor}{String with the name of the predictor variable.}
}
\value{
'ggplot' object. Bar plot with 'WOE' as the y-axis and bins of the predictor
variable as the horizontal axis.
}
\description{
Internal function within \code{create_IV()} that plots WOE graphs using an IV
object. Can also be used for plotting individual WOE graphs.
}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{is_date_format}()},
\code{\link{maxmin}()},
\code{\link{md2html}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{read_preamble}()},
\code{\link{rgb2hex}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{tstamp}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}

Other Variable Association: 
\code{\link{IV_by_period}()},
\code{\link{IV_report}()},
\code{\link{create_IV}()}

Other Information Value: 
\code{\link{IV_by_period}()},
\code{\link{IV_report}()},
\code{\link{create_IV}()}
}
\concept{Information Value}
\concept{Support}
\concept{Variable Association}
