% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert rgb to HEX code}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r, g, b}{Values that correspond to the three RGB parameters}
}
\description{
Convert rgb to HEX code
}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{is_date_format}()},
\code{\link{maxmin}()},
\code{\link{md2html}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{plot_WOE}()},
\code{\link{read_preamble}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{tstamp}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}
}
\concept{Support}
