% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals_bind.R
\name{totals_bind}
\alias{totals_bind}
\title{Row-bind an identical data frame for computing grouped totals}
\usage{
totals_bind(data, target_col, target_value = "Total")
}
\arguments{
\item{data}{data frame}

\item{target_col}{Character value of the column in which to impute "Total".
This is usually the intended grouping column.}

\item{target_value}{Character value to impute in the new data frame to row-bind.
Defaults to "Total".}
}
\description{
Row-bind an identical data frame and impute a specific
column with the \code{target_value}, which defaults as "Total". The purpose of
this is to enable to creation of summary tables with a calculated "Total"
row. See example below on usage.
}
\examples{
sq_data \%>\%
  totals_bind(target_col = "LevelDesignation", target_value = "Total") \%>\%
  collab_sum(hrvar = "LevelDesignation", return = "table")

}
