% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstamp.R
\name{tstamp}
\alias{tstamp}
\title{Generate a time stamp}
\usage{
tstamp()
}
\value{
String containing the timestamp in the format \code{'yymmdd_hhmmss'}.
}
\description{
This function generates a time stamp of the format \code{'yymmdd_hhmmss'}.
This is a support function and is not intended for direct use.
}
\seealso{
Other Support: 
\code{\link{camel_clean}()},
\code{\link{check_inputs}()},
\code{\link{combine_signals}()},
\code{\link{cut_hour}()},
\code{\link{extract_date_range}()},
\code{\link{extract_hr}()},
\code{\link{heat_colours}()},
\code{\link{is_date_format}()},
\code{\link{maxmin}()},
\code{\link{md2html}()},
\code{\link{p_test}()},
\code{\link{pairwise_count}()},
\code{\link{plot_WOE}()},
\code{\link{read_preamble}()},
\code{\link{rgb2hex}()},
\code{\link{totals_bind}()},
\code{\link{totals_col}()},
\code{\link{totals_reorder}()},
\code{\link{us_to_space}()},
\code{\link{wrap}()}
}
\concept{Support}
