% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_report.R
\name{validation_report}
\alias{validation_report}
\title{Generate a Data Validation report in HTML}
\usage{
validation_report(
  data,
  meeting_data = NULL,
  hrvar = "Organization",
  path = "validation report",
  hrvar_threshold = 150,
  timestamp = TRUE
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{meeting_data}{An optional Meeting Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "Organization"}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file extension}.}

\item{hrvar_threshold}{Numeric value determining the maximum number of unique values
to be allowed to qualify as a HR variable. This is passed directly to the \code{threshold}
argument within \code{hrvar_count_all()}.}

\item{timestamp}{Logical vector specifying whether to include a timestamp in the file name.
Defaults to TRUE.}
}
\description{
The function generates an interactive HTML report using
Standard Person Query data as an input. The report contains checks on
Workplace Analytics query outputs to provide diagnostic information
for the Analyst prior to analysis.

An additional Standard Meeting Query can be provided to perform meeting subject line
related checks. This is optional and the validation report can be run without it.
}
\details{
For your input to \code{data} or \code{meeting_data}, please use the function \code{wpa::import_wpa()}
to import your csv query files into R. This function will standardize format
and prepare the data as input for this report.

If you are passing a Ways of Working Assessment query instead of a Standard Person query
to the \code{data} argument, please also use \code{standardise_pq()} to make the variable names
consistent with a Standard Person Query.
}
\section{Checking functions within \code{validation_report()}}{

\itemize{
\item \code{check_query()}
\item \code{flag_ch_ratio()}
\item \code{hrvar_count_all()}
\item \code{identify_privacythreshold()}
\item \code{identify_nkw()}
\item \code{identify_holidayweeks()}
\item \code{subject_validate()}
\item \code{identify_tenure()}
\item \code{flag_outlooktime()}
\item \code{identify_shifts()}
\item \code{track_HR_change()}
}

You can browse each individual function for details on calculations.
}

\examples{
\dontrun{
validation_report(dv_data,
                  meeting_data = mt_data,
                  hrvar = "Organization")

}

}
\seealso{
Other Reports: 
\code{\link{IV_report}()},
\code{\link{capacity_report}()},
\code{\link{coaching_report}()},
\code{\link{collaboration_report}()},
\code{\link{connectivity_report}()},
\code{\link{generate_report}()},
\code{\link{meeting_tm_report}()},
\code{\link{subject_validate_report}()}
}
\concept{Reports}
