% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workloads_fizz.R
\name{workloads_fizz}
\alias{workloads_fizz}
\title{Distribution of Work Week Span (Fizzy Drink plot)}
\usage{
workloads_fizz(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector, defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size, defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\description{
Analyze Work Week Span distribution, and returns
a 'fizzy' scatter plot by default.
Additional options available to return a table with distribution elements.
}
\examples{
meeting_fizz(sq_data, hrvar = "Organization", return = "table")
}
\seealso{
Other Workloads: 
\code{\link{workloads_dist}()},
\code{\link{workloads_line}()},
\code{\link{workloads_summary}()},
\code{\link{workloads_trend}()}
}
\concept{Workloads}
