% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_classify.R
\name{workpatterns_classify}
\alias{workpatterns_classify}
\title{Classify working pattern personas using a rule based algorithm}
\usage{
workpatterns_classify(
  data,
  hrvar = "Organization",
  values = "percent",
  signals = "email",
  start_hour = "0900",
  end_hour = "1700",
  method = "bw",
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing data from the Hourly Collaboration query.}

\item{hrvar}{A string specifying the HR attribute to cut the data by.
Defaults to NULL. This only affects the function when "table" is returned,
and is only applicable for method = "bw".}

\item{values}{Only valid if using \code{pav} method.
Character vector to specify whether to return percentages
or absolute values in "data" and "plot". Valid values are "percent" (default)
and "abs".}

\item{signals}{Character vector to specify which collaboration metrics to use:
\itemize{
\item "email" (default) for emails only
\item "IM" for Teams messages only,
\item "unscheduled_calls" for Unscheduled Calls only
\item "meetings" for Meetings only
\item or a combination of signals, such as \code{c("email", "IM")}
}}

\item{start_hour}{A character vector specifying starting hours,
e.g. "0900". Note that this currently only supports \strong{hourly} increments.}

\item{end_hour}{A character vector specifying starting hours,
e.g. "1700". Note that this currently only supports \strong{hourly} increments.}

\item{method}{String to pass through specifying which method to use for classification.
By default, a binary week-based (bw) method is used, with options to use the
the person-average volume-based (pav) method.}

\item{return}{Character vector to specify what to return. Valid options include:
\itemize{
\item "plot": returns a heatmap plot of signal distribution by hour and archetypes (default)
\item "data": returns the raw data with the classified archetypes
\item "table": returns a summary table of the archetypes
\item "plot-area": returns an area plot of the percentages of archetypes shown over time
}}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Apply a rule based algorithm to emails or instant messages sent by hour of day.
Uses a binary week-based (bw) method by default, with options to use the
the person-average volume-based (pav) method.
}
\details{
This is a wrapper around \code{workpatterns_classify_bw()} and \code{workpatterns_classify_pav()}.
}
\examples{

# Returns a plot by default
em_data \%>\% workpatterns_classify(method = "bw")

# Return an area plot
em_data \%>\% workpatterns_classify(method = "bw", return = "plot-area")

\dontrun{

em_data \%>\% workpatterns_classify(method = "bw", return = "table")

em_data \%>\% workpatterns_classify(method = "pav")

em_data \%>\% workpatterns_classify(method = "pav", return = "plot-area")

}

}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{personas_hclust}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify_pav}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
