% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_classify_bw.R
\name{workpatterns_classify_bw}
\alias{workpatterns_classify_bw}
\title{Classify working pattern week archetypes using a rule based algorithm,
using the binary week-based (bw) method.}
\usage{
workpatterns_classify_bw(
  data,
  hrvar = NULL,
  signals = c("email", "IM"),
  start_hour = "0900",
  end_hour = "1700",
  active_threshold = 0,
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing email by hours data.}

\item{hrvar}{A string specifying the HR attribute to cut the data by.
Defaults to NULL. This only affects the function when "table" is returned.}

\item{signals}{Character vector to specify which collaboration metrics to
use:
\itemize{
\item a combination of signals, such as \code{c("email", "IM")} (default)
\item \code{"email"} for emails only
\item \code{"IM"} for Teams messages only
\item \code{"unscheduled_calls"} for Unscheduled Calls only
\item \code{"meetings"} for Meetings only
}}

\item{start_hour}{A character vector specifying start hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying finish hours,
e.g. "1700"}

\item{active_threshold}{A numeric value specifying the minimum number of
signals to be greater than in order to qualify as \emph{active}. Defaults to 0.}

\item{return}{Character vector to specify what to return. Valid options
include: - "plot": returns a heatmap plot of signal distribution by hour
and archetypes (default) - "data": returns the raw data with the classified
archetypes - "table": returns a summary table of the archetypes -
"plot-area": returns an area plot of the percentages of archetypes shown
over time}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}} Apply a rule based
algorithm to emails sent by hour of day, using the binary week-based (bw)
method.
}
\examples{
\donttest{
workpatterns_classify_bw(em_data)
}


}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{identify_shifts_wp}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_pav}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
