% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workpatterns_classify_pav.R
\name{workpatterns_classify_pav}
\alias{workpatterns_classify_pav}
\title{Classify working pattern personas using a rule based algorithm, using
the person-average volume-based (pav) method.}
\usage{
workpatterns_classify_pav(
  data,
  values = "percent",
  signals = "email",
  start_hour = "0900",
  end_hour = "1700",
  return = "plot"
)
}
\arguments{
\item{data}{A data frame containing data from the Hourly Collaboration query.}

\item{values}{Character vector to specify whether to return percentages
or absolute values in "data" and "plot". Valid values are:
\itemize{
\item "percent": percentage of signals divided by total signals (default)
\item "abs": absolute count of signals
}}

\item{signals}{Character vector to specify which collaboration metrics to use:
\itemize{
\item "email" (default) for emails only
\item "IM" for Teams messages only,
\item "unscheduled_calls" for Unscheduled Calls only
\item "meetings" for Meetings only
\item or a combination of signals, such as \code{c("email", "IM")}
}}

\item{start_hour}{A character vector specifying starting hours,
e.g. "0900"}

\item{end_hour}{A character vector specifying starting hours,
e.g. "1700"}

\item{return}{Character vector to specify what to return. Valid options include:
\itemize{
\item "plot": returns a bar plot of signal distribution by hour and archetypes (default)
\item "data": returns the raw data with the classified archetypes
\item "table": returns a summary table of the archetypes
\item "plot-area": returns an overlapping area plot
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Apply a rule based algorithm to emails or instant messages sent by hour of day.
This uses a person-average volume-based (pav) method.
}
\examples{
\donttest{
# Returns a plot by default
workpatterns_classify_pav(em_data)

# Return summary table
workpatterns_classify_pav(em_data, return = "table")

# Return an area plot
workpatterns_classify_pav(em_data, return = "plot-area")
}

}
\seealso{
Other Work Patterns: 
\code{\link{flex_index}()},
\code{\link{personas_hclust}()},
\code{\link{plot_flex_index}()},
\code{\link{workpatterns_area}()},
\code{\link{workpatterns_classify_bw}()},
\code{\link{workpatterns_classify}()},
\code{\link{workpatterns_hclust}()}
}
\concept{Work Patterns}
