% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Add a character at the start and end of a character string}
\usage{
wrap(string, wrapper = "\\"")
}
\arguments{
\item{string}{Character string to be wrapped around}

\item{wrapper}{Character to wrap around \code{string}}
}
\description{
This function adds a character at the start and end of a character
string, where the default behaviour is to add a double quote.
}
